Cell detection package and pixel size
Cell detection moved package and now uses actual requested pixel size, rather than adapting to a power of two. However if values are negative, then downsample is used instead.
