+++ qupath-core-processing/src/main/java/qupath/imagej/detect/cells/WatershedCellDetection.java
@@ -21,20 +21,21 @@
  * #L%
  */
 
-package qupath.imagej.detect.nuclei;
+package qupath.imagej.detect.cells;
 
 import java.awt.image.BufferedImage;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Iterator;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import ij.IJ;
 import ij.ImagePlus;
 import ij.Prefs;
 import ij.gui.PolygonRoi;
@@ -43,17 +44,17 @@ import ij.gui.Wand;
 import ij.measure.Calibration;
 import ij.measure.Measurements;
 import ij.plugin.filter.EDM;
-import ij.plugin.filter.MaximumFinder;
 import ij.plugin.filter.RankFilters;
 import ij.process.Blitter;
 import ij.process.ByteProcessor;
 import ij.process.ColorProcessor;
+import ij.process.FloatPolygon;
 import ij.process.FloatProcessor;
-import ij.process.FloodFiller;
 import ij.process.ImageProcessor;
 import ij.process.ImageStatistics;
 import ij.process.ShortProcessor;
 import qupath.imagej.color.ColorDeconvolutionIJ;
+import qupath.imagej.helpers.IJTools;
 import qupath.imagej.objects.PathImagePlus;
 import qupath.imagej.objects.ROIConverterIJ;
 import qupath.imagej.objects.measure.ObjectMeasurements;
@@ -67,7 +68,7 @@ import qupath.lib.analysis.algorithms.SimpleImage;
 import qupath.lib.analysis.stats.RunningStatistics;
 import qupath.lib.analysis.stats.StatisticsHelper;
 import qupath.lib.color.ColorDeconvolutionStains;
-import qupath.lib.common.GeneralTools;
+import qupath.lib.color.StainVector;
 import qupath.lib.images.ImageData;
 import qupath.lib.images.PathImage;
 import qupath.lib.images.servers.ImageServer;
@@ -76,11 +77,11 @@ import qupath.lib.measurements.MeasurementListFactory;
 import qupath.lib.measurements.MeasurementList;
 import qupath.lib.objects.PathObject;
 import qupath.lib.objects.PathObjects;
-import qupath.lib.objects.classes.PathClass;
 import qupath.lib.objects.classes.PathClassFactory;
 import qupath.lib.objects.helpers.PathObjectTools;
 import qupath.lib.plugins.AbstractTileableDetectionPlugin;
 import qupath.lib.plugins.ObjectDetector;
+import qupath.lib.plugins.parameters.DoubleParameter;
 import qupath.lib.plugins.parameters.Parameter;
 import qupath.lib.plugins.parameters.ParameterList;
 import qupath.lib.roi.PolygonROI;
@@ -89,12 +90,16 @@ import qupath.lib.roi.interfaces.PathArea;
 import qupath.lib.roi.interfaces.ROI;
 
 /**
- * Cell detection that takes into consideration membrane staining.
+ * Default command for cell detection within QuPath.
+ * <p>
+ * Assumes either a nuclear or cytoplasmic staining.
+ * <p>
+ * Quantification of membranous staining requires a separate detection command.
  * 
  * @author Pete Bankhead
  *
  */
-public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlugin<BufferedImage> {
+public class WatershedCellDetection extends AbstractTileableDetectionPlugin<BufferedImage> {
 
 	private static String[] micronParameters = {
 		"requestedPixelSizeMicrons",
@@ -116,25 +121,36 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 		"cellExpansion",
 		};
 	
+	private static String[] fluorescenceParameters = {
+			"detectionImageFluorescence"
+	};
+
+	private static String[] brightfieldParameters = {
+			"detectionImageBrightfield",
+			"maxBackground"
+	};
+
 	transient private CellDetector detector;
 	
+	private final static Logger logger = LoggerFactory.getLogger(WatershedCellDetection.class);
+	
 	static String IMAGE_OPTICAL_DENSITY = "Optical density sum";
-	static String IMAGE_HEMATOXYLIN = "Hematoxylin";
+	static String IMAGE_HEMATOXYLIN = "Hematoxylin OD";
 	
-	private ParameterList params;
+	ParameterList params;
 	
 	
-	public WatershedCellMembraneDetection() {
-		
-		Prefs.setThreads(1);
+	public WatershedCellDetection() {
 		params = new ParameterList();
 		// TODO: Use a better way to determining if pixel size is available in microns
 //		params.addEmptyParameter("detectionParameters", "Detection parameters", true);
 
-		String microns = GeneralTools.micrometerSymbol();
+		String microns = IJ.micronSymbol + "m";
 		
-		params.addEmptyParameter("paramsResolution", "Setup parameters", true);
-		
+		params.addTitleParameter("Setup parameters");
+
+		params.addIntParameter("detectionImageFluorescence", "Choose detection channel", 1, null, "Choose the channel number containing a nucleus counterstain (e.g. DAPI)");
+
 		params.addChoiceParameter("detectionImageBrightfield", "Choose detection image", IMAGE_HEMATOXYLIN, Arrays.asList(IMAGE_HEMATOXYLIN, IMAGE_OPTICAL_DENSITY),
 				"Transformed image to which to apply the detection");
 
@@ -143,7 +159,7 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 //		params.addDoubleParameter("requestedPixelSize", "Requested downsample factor", 1, "");
 
 		
-		params.addEmptyParameter("paramsNuclei", "Nucleus parameters", true);
+		params.addTitleParameter("Nucleus parameters");
 		
 		params.addDoubleParameter("backgroundRadiusMicrons", "Background radius", 8, microns, 
 				"Radius for background estimation, should be > the largest nucleus radius, or <= 0 to turn off background subtraction");
@@ -153,7 +169,7 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 				"Sigma value for Gaussian filter used to reduce noise; increasing the value stops nuclei being fragmented, but may reduce the accuracy of boundaries");
 		params.addDoubleParameter("minAreaMicrons", "Minimum area", 10, microns+"^2",
 				"Detected nuclei with an area < minimum area will be discarded");
-		params.addDoubleParameter("maxAreaMicrons", "Maximum area", 1000, microns+"^2",
+		params.addDoubleParameter("maxAreaMicrons", "Maximum area", 400, microns+"^2",
 				"Detected nuclei with an area > maximum area will be discarded");
 
 		params.addDoubleParameter("backgroundRadius", "Background radius", 15, "px", 
@@ -167,33 +183,36 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 		params.addDoubleParameter("maxArea", "Maximum area", 1000, "px^2",
 				"Detected nuclei with an area > maximum area will be discarded");
 
-		params.addDoubleParameter("threshold", "Threshold", 0.1, null, 0, 2.5,
+		params.addTitleParameter("Intensity parameters");
+		params.addDoubleParameter("threshold", "Threshold", 0.1, null,
 				"Intensity threshold - detected nuclei must have a mean intensity >= threshold");
+//		params.addDoubleParameter("threshold", "Threshold", 0.1, null, 0, 2.5,
+//				"Intensity threshold - detected nuclei must have a mean intensity >= threshold");
 		params.addDoubleParameter("maxBackground", "Max background intensity", 2, null,
 				"If background radius > 0, detected nuclei occurring on a background > max background intensity will be discarded");
 		
 //		params.addBooleanParameter("mergeAll", "Merge all", true);
 		params.addBooleanParameter("watershedPostProcess", "Split by shape", true,
 				"Split merged detected nuclei based on shape ('roundness')");
-		params.addBooleanParameter("excludeDAB", "Exclude DAB (membrane staining)", true,
+		params.addBooleanParameter("excludeDAB", "Exclude DAB (membrane staining)", false,
 				"Set to 'true' if regions of high DAB staining should not be considered nuclei; useful if DAB stains cell membranes");
 		
 		
-		params.addEmptyParameter("paramsCells", "Cell parameters", true);
+		params.addTitleParameter("Cell parameters");
 
-		params.addDoubleParameter("cellExpansionMicrons", "Cell expansion", 8, microns, 0, 25,
+		params.addDoubleParameter("cellExpansionMicrons", "Cell expansion", 5, microns, 0, 25,
 				"Amount by which to expand detected nuclei to approximate the full cell area");
-		params.addDoubleParameter("cellExpansion", "Cell expansion", 10, "px",
+		params.addDoubleParameter("cellExpansion", "Cell expansion", 5, "px",
 				"Amount by which to expand detected nuclei to approximate the full cell area");
 		
-		params.addBooleanParameter("limitExpansionByNucleusSize", "Limit cell expansion by nucleus size", false, "If checked, nuclei will not be expanded by more than their (estimated) smallest diameter in any direction - may give more realistic results for smaller, or 'thinner' nuclei");
+//		params.addBooleanParameter("limitExpansionByNucleusSize", "Limit cell expansion by nucleus size", false, "If checked, nuclei will not be expanded by more than their (estimated) smallest diameter in any direction - may give more realistic results for smaller, or 'thinner' nuclei");
 			
 		params.addBooleanParameter("includeNuclei", "Include cell nucleus", true,
 				"If cell expansion is used, optionally include/exclude the nuclei within the detected cells");
 		
 		
-		params.addEmptyParameter("paramsGeneral", "General parameters", true);
-		params.addBooleanParameter("smoothBoundaries", "Smooth boundaries", false,
+		params.addTitleParameter("General parameters");
+		params.addBooleanParameter("smoothBoundaries", "Smooth boundaries", true,
 				"Smooth the detected nucleus/cell boundaries");
 		params.addBooleanParameter("makeMeasurements", "Make measurements", true,
 				"Add default shape & intensity measurements during detection");
@@ -202,14 +221,27 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 	
 	static class CellDetector implements ObjectDetector<BufferedImage> {
 	
+		private String lastServerPath = null;
+		//private PathImage<ImagePlus> pathImage; // Caching these cause out of memory errors...
+		private ROI pathROI;
+		
 		private List<PathObject> pathObjects = null;
+//		private WatershedCellDetector detector2;
+//		private FloatProcessor fpDetection, fpH, fpDAB;
+//		private ColorDeconvolutionStains stains;
 		
 		private boolean nucleiClassified = false;
 	
 			
 		public static double getPreferredPixelSizeMicrons(ImageData<BufferedImage> imageData, ParameterList params) {
-			if (imageData.getServer().hasPixelSizeMicrons())
-				return Math.max(params.getDoubleParameterValue("requestedPixelSizeMicrons"), imageData.getServer().getAveragedPixelSizeMicrons());
+			if (imageData.getServer().hasPixelSizeMicrons()) {
+				double requestedPixelSize = params.getDoubleParameterValue("requestedPixelSizeMicrons");
+				double averagedPixelSize = imageData.getServer().getAveragedPixelSizeMicrons();
+				if (requestedPixelSize < 0)
+					requestedPixelSize = averagedPixelSize * (-requestedPixelSize);
+				requestedPixelSize = Math.max(requestedPixelSize, averagedPixelSize);
+				return requestedPixelSize;
+			}
 			return Double.NaN;
 		}
 		
@@ -222,49 +254,84 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 			// Get a PathImage if we have a new ROI
 //			boolean imageChanged = false;
 			PathImage<ImagePlus> pathImage = null;
-//			if (pathImage == null || lastServerPath == null || !lastServerPath.equals(imageData.getServerPath()) || !pathROI.equals(this.pathROI)) {
+			if (lastServerPath == null || !lastServerPath.equals(imageData.getServerPath()) || pathImage == null || !pathROI.equals(this.pathROI)) {
 				ImageServer<BufferedImage> server = imageData.getServer();
-				pathImage = PathImagePlus.createPathImage(server, pathROI, ServerTools.getDownsampleFactor(server, getPreferredPixelSizeMicrons(imageData, params), true));
-//				System.out.println("Downsample: " + pathImage.getDownsampleFactor());
-//				this.pathROI = pathROI;
-//				lastServerPath = imageData.getServerPath();
+				lastServerPath = imageData.getServerPath();
+				pathImage = PathImagePlus.createPathImage(server, pathROI, ServerTools.getDownsampleFactor(server, getPreferredPixelSizeMicrons(imageData, params), false));
+				logger.trace("Cell detection with downsample: {}", pathImage.getDownsampleFactor());
+				this.pathROI = pathROI;
 //				imageChanged = true;
-//			}
+			}
 			// Create a detector if we don't already have one for this image
 			boolean isBrightfield = imageData.isBrightfield();
-			FloatProcessor fpDetection = null, fpH = null, fpDAB = null;
-//			if (detector2 == null || imageChanged || stains != imageData.getColorDeconvolutionStains()) {
-				ImageProcessor ip = pathImage.getImage().getProcessor();
-				ColorDeconvolutionStains stains = imageData.getColorDeconvolutionStains();
-				if (ip instanceof ColorProcessor && stains != null) {
-					
-					FloatProcessor[] fps = ColorDeconvolutionIJ.colorDeconvolve((ColorProcessor)ip, stains);
-					fpH = fps[0];
-					if (stains.isH_DAB())
-						fpDAB = fps[1];
-					else
-						fpDAB = null; // At this point, only DAB is quantified (eosin ignored for H&E)
-					
-					if (!params.getParameters().get("detectionImageBrightfield").isHidden()) {
-						if (params.getChoiceParameterValue("detectionImageBrightfield").equals(IMAGE_OPTICAL_DENSITY))
-							fpDetection = ColorDeconvolutionIJ.convertToOpticalDensitySum((ColorProcessor)ip, stains.getMaxRed(), stains.getMaxGreen(), stains.getMaxBlue());
-						else
-							fpDetection = (FloatProcessor)fpH.duplicate();
-					}
-
-					
-				}
-				if (fpDetection == null) {
-					// TODO: Deal with fluorescence
-					fpDetection = ip.convertToFloatProcessor();
-					fpH = ip.convertToFloatProcessor();
-					fpDAB = null;
-				}
-//			}
+			//			if (detector2 == null || imageChanged || stains != imageData.getColorDeconvolutionStains()) {
+			//			if (imageChanged || stains != imageData.getColorDeconvolutionStains()) {
+			ImageProcessor ip = pathImage.getImage().getProcessor();
+			FloatProcessor fpDetection = null;
+			ColorDeconvolutionStains stains = imageData.getColorDeconvolutionStains();
+			Map<String, FloatProcessor> channels = new LinkedHashMap<>();
+			Map<String, FloatProcessor> channelsCell = new LinkedHashMap<>();
 			Roi roi = null;
 			if (pathROI != null)
 				roi = ROIConverterIJ.convertToIJRoi(pathROI, pathImage);
-			WatershedCellDetector detector2 = new WatershedCellDetector(fpDetection, fpH, fpDAB, roi, pathImage);
+			if (ip instanceof ColorProcessor && stains != null && isBrightfield) {
+				FloatProcessor[] fps = ColorDeconvolutionIJ.colorDeconvolve((ColorProcessor)ip, stains);
+				for (int i = 0; i < 3; i++) {
+					StainVector stain = stains.getStain(i+1);
+					if (!stain.isResidual()) {
+						channels.put(stain.getName() + " OD", fps[i]);
+						channelsCell.put(stain.getName() + " OD", fps[i]);
+					}
+				}
+//				channels.put("Hematoxylin OD",  fps[0]);
+//				if (stains.isH_DAB()) {
+//					channels.put("DAB OD", fps[1]);
+//					channelsCell.put("DAB OD", fps[1]);
+//				}
+//				else if (stains.isH_E()) {
+//					channels.put("Eosin OD", fps[1]);
+//					channelsCell.put("Eosin OD", fps[1]);
+//				}
+				
+
+				if (!params.getParameters().get("detectionImageBrightfield").isHidden()) {
+					if (params.getChoiceParameterValue("detectionImageBrightfield").equals(IMAGE_OPTICAL_DENSITY))
+						fpDetection = ColorDeconvolutionIJ.convertToOpticalDensitySum((ColorProcessor)ip, stains.getMaxRed(), stains.getMaxGreen(), stains.getMaxBlue());
+					else
+						fpDetection = (FloatProcessor)fps[0].duplicate();
+				}
+				
+				// Temporary test of the usefulness of RGB measurements...
+//				channels.put("Red", ((ColorProcessor)ip).toFloat(0, null));
+//				channels.put("Green", ((ColorProcessor)ip).toFloat(1, null));
+//				channels.put("Blue", ((ColorProcessor)ip).toFloat(2, null));
+				
+			} //else {
+			if (fpDetection == null) {
+				if (ip instanceof ColorProcessor) {
+					channels.put("Channel 1", ((ColorProcessor)ip).toFloat(0, null));
+					channels.put("Channel 2", ((ColorProcessor)ip).toFloat(1, null));
+					channels.put("Channel 3", ((ColorProcessor)ip).toFloat(2, null));
+				} else {
+					ImagePlus imp = pathImage.getImage();
+					for (int c = 1; c <= imp.getNChannels(); c++) {
+						channels.put("Channel " + c, imp.getStack().getProcessor(imp.getStackIndex(c, 0, 0)).convertToFloatProcessor());
+					}
+				}
+				// For fluorescence, measure everything
+				channelsCell.putAll(channels);
+				
+				// TODO: Deal with fluorescence... for now, defaults to first channel (may be totally wrong)
+				int detectionChannel = 1;
+				if (!isBrightfield)
+					detectionChannel = params.getIntParameterValue("detectionImageFluorescence");
+				fpDetection = channels.get("Channel " + detectionChannel);
+				if (fpDetection == null) {
+					logger.warn("Unable to find specified Channel {} - will default to Channel 1", detectionChannel);
+					fpDetection = channels.get("Channel 1");
+				}
+			}
+			WatershedCellDetector detector2 = new WatershedCellDetector(fpDetection, channels, channelsCell, roi, pathImage);
 			
 			// Create or reset the PathObjects list
 			if (pathObjects == null)
@@ -294,7 +361,7 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 			
 			detector2.runDetection(
 					backgroundRadius,
-					params.getDoubleParameterValue("maxBackground"),
+					isBrightfield ? params.getDoubleParameterValue("maxBackground") : Double.NEGATIVE_INFINITY,
 					medianRadius,
 					sigma,
 					params.getDoubleParameterValue("threshold"),
@@ -304,10 +371,12 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 					params.getBooleanParameterValue("watershedPostProcess"),
 					params.getBooleanParameterValue("excludeDAB"),
 					cellExpansion,
-					params.getBooleanParameterValue("limitExpansionByNucleusSize"),
+//					params.getBooleanParameterValue("limitExpansionByNucleusSize"),
 					params.getBooleanParameterValue("smoothBoundaries"),
 					params.getBooleanParameterValue("includeNuclei"),
-					params.getBooleanParameterValue("makeMeasurements") && isBrightfield);
+					params.getBooleanParameterValue("makeMeasurements"),
+					pathROI.getZ(),
+					pathROI.getT());// && isBrightfield);
 			
 			pathObjects.addAll(detector2.getPathObjects());
 					
@@ -347,12 +416,26 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 			map.get(name).setHidden(!pixelSizeKnown);
 		for (String name : pixelParameters)
 			map.get(name).setHidden(pixelSizeKnown);
+		
+		params.setHiddenParameters(!pixelSizeKnown, micronParameters);
+		params.setHiddenParameters(pixelSizeKnown, pixelParameters);
 
-		map.get("detectionImageBrightfield").setHidden(imageData.getColorDeconvolutionStains() == null);
+		boolean isBrightfield = imageData.isBrightfield();
+		params.setHiddenParameters(!isBrightfield, brightfieldParameters);
+		params.setHiddenParameters(isBrightfield, fluorescenceParameters);
+		
+		if (!isBrightfield) {
+			if (imageData.getServer().getBitsPerPixel() > 8)
+				((DoubleParameter)params.getParameters().get("threshold")).setValue(100.0);
+			else
+				((DoubleParameter)params.getParameters().get("threshold")).setValue(25.0);
+		}
+
+//		map.get("detectionImageBrightfield").setHidden(imageData.getColorDeconvolutionStains() == null);
 
 		map.get("excludeDAB").setHidden(imageData.getColorDeconvolutionStains() == null || !imageData.getColorDeconvolutionStains().isH_DAB());
 		
-		map.get("makeMeasurements").setHidden(!imageData.isBrightfield());
+//		map.get("makeMeasurements").setHidden(!imageData.isBrightfield());
 
 		return params;
 	}
@@ -376,13 +459,14 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 	
 	static class WatershedCellDetector {
 		
-		final private static Logger logger = LoggerFactory.getLogger(WatershedCellDetector.class);
 		
 		private boolean refineBoundary = true; // TODO: Consider making this variable accessible
 		
 		
 		private double backgroundRadius = 15;
 		private double maxBackground = 0.3;
+		
+		private int z = 0, t = 0;
 		
 		private boolean lastRunCompleted = false;
 		
@@ -400,16 +484,15 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 		private boolean excludeDAB = false;
 		private boolean smoothBoundaries = false;
 
-		private boolean limitExpansionByNucleusSize = false;
+//		private boolean limitExpansionByNucleusSize = false;
 
 		private boolean makeMeasurements = true;
 		
 		private Roi roi = null;
 		private FloatProcessor fpDetection = null;
-		private FloatProcessor fpH = null;
-		private FloatProcessor fpDAB = null;
+		private Map<String, FloatProcessor> channels = new LinkedHashMap<>(); // Map of channels to measure for nuclei only, and their names
+		private Map<String, FloatProcessor> channelsCell = new LinkedHashMap<>(); // Map of channels to measure for cell/cytoplasm, and their names
 		private ImageProcessor ipToMeasure = null;
-		private ImageProcessor ipBackground = null;
 		private List<PolygonRoi> rois = null;
 		private ByteProcessor bpLoG = null;
 		
@@ -418,13 +501,15 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 		
 		private PathImage<ImagePlus> pathImage = null;
 		
-		public WatershedCellDetector(FloatProcessor fpDetection, FloatProcessor fpH, FloatProcessor fpDAB, Roi roi, PathImage<ImagePlus> pathImage) {
+		public WatershedCellDetector(FloatProcessor fpDetection, Map<String, FloatProcessor> channels, Map<String, FloatProcessor> channelsCell, Roi roi, PathImage<ImagePlus> pathImage) {
 			this.fpDetection = fpDetection;
-			this.fpH = fpH;
-			this.fpDAB = fpDAB;
+			if (channels != null)
+				this.channels.putAll(channels);
+			if (channelsCell != null)
+				this.channelsCell.putAll(channelsCell);
 			this.roi = roi;
 			this.pathImage = pathImage;
-//			Prefs.setThreads(1);
+			Prefs.setThreads(1);
 		}
 		
 		
@@ -454,10 +539,7 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 						if (bpMask.getf(i) != 0f) {
 							ipBackground.setf(i, Float.NEGATIVE_INFINITY);
 						}
-					}				
-//				} else {
-//					// Don't return a mask - all pixels are ok
-//					System.out.println("Skipping background mask!");
+					}
 				}
 			}
 			
@@ -484,63 +566,34 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 				// Use Laplacian of Gaussian filtering followed by watershed transform to determine possible nucleus segments
 				// Result will be a dramatic over-segmentation...
 				FloatProcessor fpLoG = (FloatProcessor)fpDetection.duplicate();
-								
-////				fpLoG.copyBits(fpDAB, 0, 0, Blitter.ADD); // Testing is adding the optical densities helps...
-//				// Check we have some above-threshold pixels - if not, don't do more
-//				int indAbove = 0;
-//				int n = fpLoG.getWidth() * fpLoG.getHeight();
-//				while (indAbove < n && fpLoG.getf(indAbove) < threshold)
-//					indAbove++;
-//				if (indAbove == n)
-//					return;
 
 				// Start off with a median filter to reduce texture, if necessary
 				RankFilters rf = new RankFilters();
 				if (medianRadius > 0)
 					rf.rank(fpLoG, medianRadius, RankFilters.MEDIAN);
 
-				// Subtract background first, if needed
-				if (backgroundRadius > 0) {
-					
-//					// Could mask out definitely-background pixels if we wanted (and could be confident in background radius)... although it doesn't actually help all that much
-//					ByteProcessor bpPossible = SimpleThresholding.thresholdAboveEquals(fpH, (float)threshold);
-//					rf.rank(bpPossible, backgroundRadius, RankFilters.MAX);
-//					if (roi != null)
-//						ROILabeling.fillOutside(bpPossible, roi, 0);
-//					for (int i = 0; i < width * height; i++) {
-//						if (bpPossible.get(i) == 0)
-//							fpLoG.setf(i, Float.NEGATIVE_INFINITY);
-//					}
-////					bpPossible.max(1);
-////					new ImagePlus("Possible", bpPossible.duplicate()).show();
-
-//					ROILabeling.fillOutside(fpLoG, roi, Float.NEGATIVE_INFINITY);
-
-					ipBackground = fpLoG.duplicate();
-									
-					bpBackgroundMask = limitedOpeningByReconstruction(fpLoG, ipBackground, backgroundRadius, maxBackground);
-					
-//					ipBackground = MorphologicalReconstructionInteger.openingByReconstruction(fpLoG, backgroundRadius);
-					fpLoG.copyBits(ipBackground, 0, 0, Blitter.SUBTRACT);
-					ipToMeasure = fpLoG.duplicate();
-//					new ImagePlus("Background", ipBackground.duplicate()).show();
-				} else {
-					ipToMeasure = fpDetection;
-					ipBackground = null;
-				}
-				
 				//--------NEW--------
-				if (excludeDAB && fpH != null && fpDAB != null) {
+				if (excludeDAB && channels.containsKey("Hematoxylin OD") && channels.containsKey("DAB OD")) {
 					// If we are avoiding DAB, set pixels away from potential nuclei to zero
+					FloatProcessor fpDAB = channels.get("DAB OD");
 					fpDAB.setRoi(roi);
-					ByteProcessor bpH = SimpleThresholding.greaterThanOrEqual(fpH, fpDAB);
+					ByteProcessor bpH = SimpleThresholding.greaterThanOrEqual(channels.get("Hematoxylin OD"), fpDAB);
 					bpH.multiply(1.0/255.0);
-					rf.rank(bpH, 2.5, RankFilters.MEDIAN); // TODO: Check hard-coded filter sizes for reasonableness
+					rf.rank(bpH, 2.5, RankFilters.MEDIAN);
 					rf.rank(bpH, 2.5, RankFilters.MAX);
 					fpLoG.copyBits(bpH, 0, 0, Blitter.MULTIPLY);
 				}
-//				new ImagePlus("Log", fpLoG.duplicate()).show();
 				//--------END_NEW--------
+				
+				// Subtract background first, if needed
+				if (backgroundRadius > 0) {
+					ImageProcessor ipBackground = fpLoG.duplicate();
+					bpBackgroundMask = limitedOpeningByReconstruction(fpLoG, ipBackground, backgroundRadius, maxBackground);
+					fpLoG.copyBits(ipBackground, 0, 0, Blitter.SUBTRACT);
+					ipToMeasure = fpLoG.duplicate();
+				} else {
+					ipToMeasure = fpDetection;
+				}
 				
 				// Apply (approximation of) Laplacian of Gaussian filter
 				fpLoG.blurGaussian(sigma);
@@ -549,30 +602,31 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 				// Threshold the main LoG image
 				bpLoG = SimpleThresholding.thresholdAbove(fpLoG, 0f);
 				// Need to set the threshold very slightly above zero for ImageJ
+				// TODO: DECIDE ON USING MY WATERSHED OR IMAGEJ'S....
 				fpLoG.setRoi(roi);
-				
-//				ROILabeling.fillOutside(fpLoG, roi, Float.NEGATIVE_INFINITY);
 				
 				ImageProcessor ipTemp = RegionalExtrema.findRegionalMaxima(fpLoG, 0.001f, false);
 				ImageProcessor ipLabels = ROILabeling.labelImage(ipTemp, 0, false);
-//				new ImagePlus("Labels before", ipLabels.duplicate()).show();
 				Watershed.doWatershed(fpLoG, ipLabels, 0, false);
-//				new ImagePlus("Labels after", ipLabels.duplicate()).show();
-				
 				
 				ipLabels.setThreshold(0.5, Double.POSITIVE_INFINITY, ImageProcessor.NO_LUT_UPDATE);
 				// TODO: Consider 4/8 connectivity for watershed nucleus ROIs
 				rois = ROILabeling.getFilledPolygonROIs(ipLabels, Wand.FOUR_CONNECTED);			
-//				rois = ROILabeling.getFilledPolygonROIsExperimental(ipLabels);
 				
-				
-//				new ImagePlus("Labels", ipLabels.duplicate()).show();
 				if (Thread.currentThread().isInterrupted())
 					return;
 			} 
 			
 			if (bp == null)
-				bp = new ByteProcessor(width, height);			
+				bp = new ByteProcessor(width, height);	
+			
+//			// TODO: Consider application of an automated threshold
+//			if (threshold < 0) {
+//				ipToMeasure.resetRoi();
+//				ImageStatistics stats = ipToMeasure.getStatistics();
+//				threshold = stats.mean;// + stats.stdDev;
+//				logger.info("Mean threshold set: " + threshold);
+//			}
 
 			bp.setValue(255);
 			for (Roi r : rois) {
@@ -588,16 +642,9 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 					if (bpBackgroundMask.getStatistics().mean > 0)
 						continue;				
 				}
-//				if (ipBackground != null && !Double.isNaN(maxBackground)) {
-//					ipBackground.setRoi(r);
-//					if (ipBackground.getStatistics().mean > maxBackground)
-//						continue;
-//				}
 				// Fill the ROI to keep it
 				bp.fill(r);
 			}
-			
-//			new ImagePlus("BP early", bp.duplicate()).show();
 			
 			if (Thread.currentThread().isInterrupted())
 				return;
@@ -619,34 +666,31 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 			// TODO: Look at the better boundary clearing implemented in Fast_nucleus_counts
 			if (roi != null)
 				ROILabeling.clearOutside(bp, roi);
-//				ROILabeling.clearBoundary(bp, roi, 0);
 			
 			// Locate nucleus ROIs
 			bp.setThreshold(127, Double.POSITIVE_INFINITY, ImageProcessor.NO_LUT_UPDATE);
 			
 			
-//			new ImagePlus("BP later", bp.duplicate()).show();
+			if (IJ.debugMode) {
+				IJTools.quickShowImage("Binary", bp.duplicate());
+			}
 
 			//----------------------------
 			// MINOR BOUNDARY REFINEMENT
 			// The idea is that Gaussian smoothing tends to cause the boundaries of 'thin' nuclei to be overestimated;
 			// this uses a smaller filter to correct instances where the boundary has moved by just one pixel
 			if (refineBoundary && sigma > 1.5) {
-				
-//				new ImagePlus("Before", bp.duplicate()).show();
-				
-				FloatProcessor fpLoG = (FloatProcessor)fpDetection.duplicate();
-				fpLoG.blurGaussian(1);
-				fpLoG.convolve(new float[]{0, -1, 0, -1, 4, -1, 0, -1, 0}, 3, 3);
-				ByteProcessor bp2 = SimpleThresholding.thresholdAbove(fpLoG, 0f);
+				FloatProcessor fpBoundaryCleanup = (FloatProcessor)fpDetection.duplicate();
+				fpBoundaryCleanup.blurGaussian(1);
+				fpBoundaryCleanup.convolve(new float[]{0, -1, 0, -1, 4, -1, 0, -1, 0}, 3, 3);
+				ByteProcessor bp2 = SimpleThresholding.thresholdAbove(fpBoundaryCleanup, 0f);
 				bp2.copyBits(bp, 0, 0, Blitter.MIN); // Remove everything not detected in bp
 				bp.filter(ByteProcessor.MIN);
 				bp.copyBits(bp2, 0, 0, Blitter.MAX);
-
-//				new ImagePlus("After", bp.duplicate()).show();
-
 				regenerateROIs = true;
 			}
+			
+			//----------------------------
 			
 			roisNuclei = ROILabeling.getFilledPolygonROIs(bp, Wand.FOUR_CONNECTED);
 
@@ -659,14 +703,15 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 				Iterator<PolygonRoi> iter = roisNuclei.iterator();
 				while (iter.hasNext()) {
 					Roi roiTemp = iter.next();
-					fpDetection.setRoi(roiTemp);
-					double area = ImageStatistics.getStatistics(fpDetection, Measurements.AREA, null).pixelCount;
-					if ((minArea > 0 && area < minArea) || (maxArea > 0 && area > maxArea)) {
+					ipToMeasure.setRoi(roiTemp);
+					ImageStatistics stats = ImageStatistics.getStatistics(ipToMeasure, Measurements.AREA | Measurements.MEAN, null);
+					double area = stats.pixelCount;
+					if ((stats.mean < threshold) || (minArea > 0 && area < minArea) || (maxArea > 0 && area > maxArea)) {
 						iter.remove();
 						bp.fill(roiTemp);
 					}
 				}
-				fpDetection.resetRoi();
+				ipToMeasure.resetRoi();
 			}
 			
 			
@@ -674,98 +719,14 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 			ShortProcessor ipLabels = new ShortProcessor(width, height);
 			ROILabeling.labelROIs(ipLabels, roisNuclei);
 			
-			
-			
-			
-			//----------------------------
-			Map<Float, PolygonRoi> roisCellsList = null;
-			ImageProcessor ipLabelsCells = null;
-			int nCells = 0;
-			
-			// Membrane detection (new 9/8/2015)
-			if (excludeDAB && fpDAB != null) {
-				
-				FloatProcessor fpMembranes = (FloatProcessor)fpDAB.duplicate();
-				fpMembranes.blurGaussian(2); // TODO: Enable a different sigma parameter?
-				
-				float membraneThreshold = .2f; // TODO: Enable a different membrane threshold?
-				for (int i = 0; i < width * height; i++) {
-					fpMembranes.setf(i, membraneThreshold - fpMembranes.getf(i));
-				}
-				fpMembranes.max(0.);
-				
-				// Fill nuclei with max values
-				fpMembranes.setValue(0.1);
-				for (Roi r : roisNuclei) {
-					fpMembranes.fill(r);
-				}
-				
-				// Locate regional maxima that will be used for detection
-				ByteProcessor bpMarkers = new MaximumFinder().findMaxima(fpMembranes, 0.09, ImageProcessor.NO_THRESHOLD, MaximumFinder.IN_TOLERANCE, false, false);
-				
-				// Remove any impossibly-large areas
-				ROILabeling.removeByAreas(bpMarkers, 1, maxArea, true);
-				
-				// Determine a mask of the furthest anything can expand
-				ByteProcessor bpMaxExpansion = (ByteProcessor)bpMarkers.duplicate();
-				new RankFilters().rank(bpMaxExpansion, cellExpansion, RankFilters.MAX);
-				
-				// Update the mask to remove completely unstained pixels
-				for (int i = 0; i < width * height; i++) {
-					if (fpH.getf(i) + fpDAB.getf(i) < 0.025f)
-						bpMaxExpansion.set(i, 0);
-				}
-				
-				// Expand all potential cells as much as possible
-				float minThreshold = (float)(fpMembranes.getStatistics().min - 0.05);
-				for (int i = 0; i < width * height; i++) {
-					if (bpMaxExpansion.get(i) == 0)
-						fpMembranes.setf(i, minThreshold);
-				}
-				fpMembranes.resetMinAndMax();
-				int lastLabel = roisNuclei.size();
-				ipLabelsCells = ipLabels.duplicate();
-				for (int i = 0; i < width * height; i++) {
-					if (bpMarkers.get(i) != 0 && ipLabelsCells.get(i) == 0) {
-						ipLabelsCells.set(i, Short.MAX_VALUE);
-					}
-				}
-				FloodFiller ff = new FloodFiller(ipLabelsCells);
-				for (int i = 0; i < width * height; i++) {
-					if (ipLabelsCells.get(i) == Short.MAX_VALUE) {
-						lastLabel++;
-						ipLabelsCells.setValue(lastLabel);
-						ff.fill(i%width, i/width);
-					}
-				}
-				
-//				fpMembranes.resetMinAndMax();
-//				new ImagePlus("BEFORE", fpMembranes.duplicate()).show();
-//				ipLabels = ROILabeling.labelImage(bpMarkers, false);
-				Watershed.doWatershed(fpMembranes, ipLabelsCells, minThreshold+.025, false);
-				ipLabelsCells.setThreshold(0.5, Double.POSITIVE_INFINITY, ImageProcessor.NO_LUT_UPDATE);
-				roisCellsList = ROILabeling.getFilledPolygonROIsFromLabels(ipLabelsCells, Wand.FOUR_CONNECTED);
-				nCells = lastLabel;
-			}
-			
-			
-			
-			//----------------------------
-			
-			
-			
-			
-			
-			// Measure nuclei
-			List<RunningStatistics> statsHematoxylin = null;
-			List<RunningStatistics> statsDAB = null;
+			// Measure nuclei for all required channels
+			Map<String, List<RunningStatistics>> statsMap = new LinkedHashMap<>();
 			if (makeMeasurements) {
-				statsHematoxylin = StatisticsHelper.createRunningStatisticsList(roisNuclei.size());
 				SimpleImage imgLabels = new PixelImageIJ(ipLabels);
-				StatisticsHelper.computeRunningStatistics(new PixelImageIJ(fpH), imgLabels, statsHematoxylin);
-				if (fpDAB != null) {
-					statsDAB = StatisticsHelper.createRunningStatisticsList(roisNuclei.size());
-					StatisticsHelper.computeRunningStatistics(new PixelImageIJ(fpDAB), imgLabels, statsDAB);
+				for (String key : channels.keySet()) {
+					List<RunningStatistics> statsList = StatisticsHelper.createRunningStatisticsList(roisNuclei.size());
+					StatisticsHelper.computeRunningStatistics(new PixelImageIJ(channels.get(key)), imgLabels, statsList);
+					statsMap.put(key, statsList);
 				}
 			}
 			
@@ -778,49 +739,37 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 			List<PathObject> nucleiObjects = new ArrayList<>();
 			Calibration cal = pathImage.getImage().getCalibration();
 			for (int i = 0; i < roisNuclei.size(); i++) {
-				PolygonRoi r = roisNuclei.get(i);
+				PolygonRoi rOrig = roisNuclei.get(i);
 				
+				PolygonRoi r = rOrig;
 				if (smoothBoundaries)
-					r = new PolygonRoi(r.getInterpolatedPolygon(Math.min(2.5, r.getNCoordinates()*0.1), true), Roi.POLYGON);
+					r = new PolygonRoi(rOrig.getInterpolatedPolygon(Math.min(2.5, rOrig.getNCoordinates()*0.1), true), Roi.POLYGON);
 				
-				PolygonROI pathROI = ROIConverterIJ.convertToPolygonROI(r, cal, pathImage.getDownsampleFactor());
+				PolygonROI pathROI = ROIConverterIJ.convertToPolygonROI(r, cal, pathImage.getDownsampleFactor(), 0, z, t);
 				
 				if (smoothBoundaries) {
-//					int nBefore = pathROI.nVertices();
 					pathROI = ShapeSimplifier.simplifyPolygon(pathROI, pathImage.getDownsampleFactor()/4.0);
-//					int nAfter = pathROI.nVertices();
-//					System.out.println("Vertices removed: " + (nBefore - nAfter));
 				}
 				
 				// Create a new shared measurement list
 				MeasurementList measurementList = MeasurementListFactory.createMeasurementList(makeMeasurements ? 30 : 0, MeasurementList.TYPE.FLOAT);
 				
 				if (makeMeasurements) {
-					ObjectMeasurements.addShapeStatistics(measurementList, r, fpH, cal, "Nucleus: ");
+					ObjectMeasurements.addShapeStatistics(measurementList, r, fpDetection, cal, "Nucleus: ");
 	
-	//				PathObject pathObject = new PathDetectionObject(pathROI, PathPrefs.getNucleusClass());
-	//				PathObjectIJ.computeShapeStatistics(pathObject, pathImage, fpH, pathImage.getImage().getCalibration(), "Nucleus: ");
-					
-					RunningStatistics stats = statsHematoxylin.get(i);
-					measurementList.addMeasurement("Nucleus: Hematoxylin OD mean", stats.getMean());
-					measurementList.addMeasurement("Nucleus: Hematoxylin OD sum", stats.getSum());
-					measurementList.addMeasurement("Nucleus: Hematoxylin OD std dev", stats.getStdDev());
-					measurementList.addMeasurement("Nucleus: Hematoxylin OD max", stats.getMax());
-					measurementList.addMeasurement("Nucleus: Hematoxylin OD min", stats.getMin());
-					measurementList.addMeasurement("Nucleus: Hematoxylin OD range", stats.getRange());
-					if (statsDAB != null) {
-						stats = statsDAB.get(i);
-						measurementList.addMeasurement("Nucleus: DAB OD mean", stats.getMean());
-						measurementList.addMeasurement("Nucleus: DAB OD sum", stats.getSum());
-						measurementList.addMeasurement("Nucleus: DAB OD std dev", stats.getStdDev());
-						measurementList.addMeasurement("Nucleus: DAB OD max", stats.getMax());
-						measurementList.addMeasurement("Nucleus: DAB OD min", stats.getMin());
-						measurementList.addMeasurement("Nucleus: DAB OD range", stats.getRange());
+					for (String key : channels.keySet()) {
+						List<RunningStatistics> statsList = statsMap.get(key);
+						RunningStatistics stats = statsList.get(i);
+						measurementList.addMeasurement("Nucleus: " + key + " mean", stats.getMean());
+						measurementList.addMeasurement("Nucleus: " + key + " sum", stats.getSum());
+						measurementList.addMeasurement("Nucleus: " + key + " std dev", stats.getStdDev());
+						measurementList.addMeasurement("Nucleus: " + key + " max", stats.getMax());
+						measurementList.addMeasurement("Nucleus: " + key + " min", stats.getMin());
+						measurementList.addMeasurement("Nucleus: " + key + " range", stats.getRange());
 					}
 				}
 				
 				// TODO: It would be more efficient to measure the hematoxylin intensities along with the shapes
-//				PathObject pathObject = new PathDetectionObject(pathROI, PathClassFactory.getNucleusClass(), measurementList);
 				PathObject pathObject = PathObjects.createDetectionObject(pathROI, null, measurementList);
 				nucleiObjects.add(pathObject);
 
@@ -829,169 +778,101 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 			if (Thread.currentThread().isInterrupted())
 				return;
 
-//			List<Roi> roisCellsList = null;
+			List<Roi> roisCellsList = null;
 			
 			// Optionally expand the nuclei to become cells
 			if (cellExpansion > 0) {
+				FloatProcessor fpEDM = new EDM().makeFloatEDM(bp, (byte)255, false);
+				fpEDM.multiply(-1);
 				
-				Map<Float, PolygonRoi> roisCells;
-				if (roisCellsList == null) {
+				double cellExpansionThreshold = -cellExpansion;
 				
-					FloatProcessor fpEDM = new EDM().makeFloatEDM(bp, (byte)255, false);
-					fpEDM.multiply(-1);
-					
-					// ---------------------- EXPERIMENTAL - CONSTRAIN CELL EXPANSION BY NUCLEUS SIZE
-					double cellExpansionThreshold = -cellExpansion;
-					if (limitExpansionByNucleusSize) {
-						MaximumFinder mf = new MaximumFinder();
-						ByteProcessor bpVoronoi = mf.findMaxima(fpEDM, .5, cellExpansionThreshold, MaximumFinder.SEGMENTED, false, false);
-						FloatProcessor fpEDM2 = new EDM().makeFloatEDM(bp, (byte)0, false);
-						
-						FloatProcessor fpMarkers = new FloatProcessor(fpEDM2.getWidth(), fpEDM2.getHeight());
-						Arrays.fill((float[])(fpMarkers.getPixels()), Float.NEGATIVE_INFINITY);
-						for (Roi r : roisNuclei) {
-							fpEDM2.setRoi(r);
-							double max = fpEDM2.getStatistics().max;
-							fpMarkers.setValue(-2. * max);
-							fpMarkers.fill(r);
-						}
-						
-						for (int i = 0; i < bpVoronoi.getWidth()*bpVoronoi.getHeight(); i++) {
-							if (bpVoronoi.getf(i) == 0)
-								fpEDM.setf(i, Float.NEGATIVE_INFINITY);
-						}
-						if (!MorphologicalReconstruction.morphologicalReconstruction(fpMarkers, fpEDM))
-							logger.error("Problem during morphological reconstruction!");
-						ByteProcessor bpMax = mf.findMaxima(fpMarkers, 0.0001, MaximumFinder.IN_TOLERANCE, false);
-						fpEDM.copyBits(bpMax, 0, 0, Blitter.COPY);
-						cellExpansionThreshold = 128; // Cell expansion has already been applied
-					}
-					
-					// ---------------------- END EXPERIMENTAL
-					
-					// Create cell ROIs
-					ipLabelsCells = ipLabels.duplicate();
-					Watershed.doWatershed(fpEDM, ipLabelsCells, cellExpansionThreshold, false);
-					roisCells = ROILabeling.getFilledPolygonROIsFromLabels(ipLabelsCells, roisNuclei.size());
-					nCells = roisCells.size();
-				} else {
-					roisCells = roisCellsList;
-				}
-				
-				
-				
+				// Create cell ROIs
+				ImageProcessor ipLabelsCells = ipLabels.duplicate();
+				Watershed.doWatershed(fpEDM, ipLabelsCells, cellExpansionThreshold, false);
+				PolygonRoi[] roisCells = ROILabeling.labelsToFilledROIs(ipLabelsCells, roisNuclei.size());
 				
 				// Compute cell DAB stats
-				List<RunningStatistics> statsDABCell = null;
-				if (fpDAB != null && makeMeasurements) {
-					statsDABCell = StatisticsHelper.createRunningStatisticsList(nCells);
-					StatisticsHelper.computeRunningStatistics(new PixelImageIJ(fpDAB), new PixelImageIJ(ipLabelsCells), statsDABCell);
+				Map<String, List<RunningStatistics>> statsMapCell = new LinkedHashMap<>();
+				if (makeMeasurements) {
+					for (String key : channelsCell.keySet()) {
+						List<RunningStatistics> statsList = StatisticsHelper.createRunningStatisticsList(roisNuclei.size());
+						StatisticsHelper.computeRunningStatistics(new PixelImageIJ(channelsCell.get(key)), new PixelImageIJ(ipLabelsCells), statsList);
+						statsMapCell.put(key, statsList);
+					}
 				}
 							
 				// Create labelled image for cytoplasm, i.e. remove all nucleus pixels
-				// TODO: Make a buffer zone between nucleus and cytoplasm?!
+				// TODO: Make a buffer zone between nucleus and cytoplasm!
 				for (int i = 0; i < ipLabels.getWidth() * ipLabels.getHeight(); i++) {
 					if (ipLabels.getf(i) != 0)
 						ipLabelsCells.setf(i, 0f);
 				}
 				
 				// Compute cytoplasm stats
-				List<RunningStatistics> statsDABCytoplasm = null;
-				if (includeNuclei && fpDAB != null && makeMeasurements) {
-					statsDABCytoplasm = StatisticsHelper.createRunningStatisticsList(nCells);
-					StatisticsHelper.computeRunningStatistics(new PixelImageIJ(fpDAB), new PixelImageIJ(ipLabelsCells), statsDABCytoplasm);
-				}
-				
-				// Create membrane stats
-				List<RunningStatistics> statsDABMembrane = null;
-				if (includeNuclei && excludeDAB && fpDAB != null && makeMeasurements) {
-					statsDABMembrane = StatisticsHelper.createRunningStatisticsList(nCells);
-					ImageProcessor ipLabelsMembrane = new ShortProcessor(width, height);
-					// TODO: WARNING!  This method of creating measurements doesn't permit membranes to overlap.
-					// This means that sometimes one cell could 'steal' a bit of the membrane of another cell.
-					// However it's unlikely this actually make any substantial difference...
-					for (Entry<Float, PolygonRoi> entry : roisCells.entrySet()) {
-						Float label = entry.getKey();
-						PolygonRoi roiTemp = entry.getValue();
-						ipLabelsMembrane.setValue(label.doubleValue());
-						ipLabelsMembrane.draw(roiTemp);
+				Map<String, List<RunningStatistics>> statsMapCytoplasm = new LinkedHashMap<>();
+				if (makeMeasurements) {
+					for (String key : channelsCell.keySet()) {
+						List<RunningStatistics> statsList = StatisticsHelper.createRunningStatisticsList(roisNuclei.size());
+						StatisticsHelper.computeRunningStatistics(new PixelImageIJ(channelsCell.get(key)), new PixelImageIJ(ipLabelsCells), statsList);
+						statsMapCytoplasm.put(key, statsList);
 					}
-					StatisticsHelper.computeRunningStatistics(new PixelImageIJ(fpDAB), new PixelImageIJ(ipLabelsMembrane), statsDABMembrane);
-//					ipLabelsMembrane.resetMinAndMax();
-//					new ImagePlus("LABELS", ipLabelsMembrane.duplicate()).show();
 				}
 				
 				
 				// Create cell objects
-//				PathClass cellClass = PathClassFactory.getPathClass("Cell", ColorTools.makeRGB(255, 200, 0));
-				PathClass cellClass = null;
-				for (Entry<Float, PolygonRoi> entry : roisCells.entrySet()) {
-					PolygonRoi r = entry.getValue();
+				roisCellsList = new ArrayList<>(roisCells.length); // In case we need texture measurements, store all cell ROIs
+				for (int i = 0; i < roisCells.length; i++) {
+					PolygonRoi r = roisCells[i];
 					if (r == null)
 						continue;
 					if (smoothBoundaries)
 						r = new PolygonRoi(r.getInterpolatedPolygon(Math.min(2.5, r.getNCoordinates()*0.1), false), Roi.POLYGON); // TODO: Check this smoothing - it can be troublesome, causing nuclei to be outside cells
 //						r = smoothPolygonRoi(r);
 
-					PolygonROI pathROI = ROIConverterIJ.convertToPolygonROI(r, pathImage.getImage().getCalibration(), pathImage.getDownsampleFactor());
+					PolygonROI pathROI = ROIConverterIJ.convertToPolygonROI(r, pathImage.getImage().getCalibration(), pathImage.getDownsampleFactor(), 0, z, t);
 					if (smoothBoundaries)
 						pathROI = ShapeSimplifier.simplifyPolygon(pathROI, pathImage.getDownsampleFactor()/4.0);
 
 					
 					MeasurementList measurementList = null;
 					PathObject nucleus = null;
-					int label = entry.getKey().intValue();
-					if (label < nucleiObjects.size())
-						nucleus = nucleiObjects.get(label-1);
-
-					// If we don't have a nucleus, check the cell area isn't too large
-					if (nucleus == null && pathROI.getArea() > maxArea * 2)
-						continue;
-
-					// Prepare measurement list with/without nucleus
-					if (includeNuclei && nucleus != null) {
+					if (includeNuclei) {
 						// Use the nucleus' measurement list
+						nucleus = nucleiObjects.get(i);
 						measurementList = nucleus.getMeasurementList();					
 					} else {
 						// Create a new measurement list
 						measurementList = MeasurementListFactory.createMeasurementList(makeMeasurements ? 12 : 0, MeasurementList.TYPE.GENERAL);
-						nucleus = null;
 					}
-					
-					
+									
 					// Add cell shape measurements
 					if (makeMeasurements) {
 						ObjectMeasurements.addShapeStatistics(measurementList, r, fpDetection, pathImage.getImage().getCalibration(), "Cell: ");
 	//					ObjectMeasurements.computeShapeStatistics(pathObject, pathImage, fpH, pathImage.getImage().getCalibration());
 	
 						// Add cell measurements
-						if (statsDABCell != null) {
-							RunningStatistics stats = statsDABCell.get(label-1);
-							measurementList.addMeasurement("Cell: DAB OD mean", stats.getMean());
-							measurementList.addMeasurement("Cell: DAB OD std dev", stats.getStdDev());
-							measurementList.addMeasurement("Cell: DAB OD max", stats.getMax());
-							measurementList.addMeasurement("Cell: DAB OD min", stats.getMin());
-	//						pathObject.addMeasurement("Cytoplasm: DAB OD range", stats.getRange());
+						for (String key : channelsCell.keySet()) {
+							if (statsMapCell.containsKey(key)) {
+								RunningStatistics stats = statsMapCell.get(key).get(i);
+								measurementList.addMeasurement("Cell: " + key + " mean", stats.getMean());
+								measurementList.addMeasurement("Cell: " + key + " std dev", stats.getStdDev());
+								measurementList.addMeasurement("Cell: " + key + " max", stats.getMax());
+								measurementList.addMeasurement("Cell: " + key + " min", stats.getMin());
+		//						pathObject.addMeasurement("Cytoplasm: " + key + " range", stats.getRange());
+							}
 						}
-						
-						// Add cytoplasm measurements
-						if (statsDABCytoplasm != null) {
-							RunningStatistics stats = statsDABCytoplasm.get(label-1);
-							measurementList.addMeasurement("Cytoplasm: DAB OD mean", stats.getMean());
-							measurementList.addMeasurement("Cytoplasm: DAB OD std dev", stats.getStdDev());
-							measurementList.addMeasurement("Cytoplasm: DAB OD max", stats.getMax());
-							measurementList.addMeasurement("Cytoplasm: DAB OD min", stats.getMin());
-	//						pathObject.addMeasurement("Cytoplasm: DAB OD range", stats.getRange());
-						}
-						
-						// Add membrane measurements
-						if (statsDABMembrane != null) {
-							RunningStatistics stats = statsDABMembrane.get(label-1);
-							measurementList.addMeasurement("Membrane: DAB OD mean", stats.getMean());
-							measurementList.addMeasurement("Membrane: DAB OD std dev", stats.getStdDev());
-							measurementList.addMeasurement("Membrane: DAB OD max", stats.getMax());
-							measurementList.addMeasurement("Membrane: DAB OD min", stats.getMin());
-	//						pathObject.addMeasurement("Cytoplasm: DAB OD range", stats.getRange());
+							
+							// Add cytoplasm measurements
+						for (String key : channelsCell.keySet()) {
+							if (statsMapCytoplasm.containsKey(key)) {
+								RunningStatistics stats = statsMapCytoplasm.get(key).get(i);
+								measurementList.addMeasurement("Cytoplasm: " + key + " mean", stats.getMean());
+								measurementList.addMeasurement("Cytoplasm: " + key + " std dev", stats.getStdDev());
+								measurementList.addMeasurement("Cytoplasm: " + key + " max", stats.getMax());
+								measurementList.addMeasurement("Cytoplasm: " + key + " min", stats.getMin());
+		//						pathObject.addMeasurement("Cytoplasm: " + key + " range", stats.getRange());
+							}
 						}
 						
 						// Add nucleus area ratio, if available
@@ -1005,8 +886,10 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 
 					
 					// Create & store the cell object
-					PathObject pathObject = PathObjects.createCellObject(pathROI, nucleus == null ? null : nucleus.getROI(), cellClass, measurementList);
+					PathObject pathObject = PathObjects.createCellObject(pathROI, nucleus == null ? null : nucleus.getROI(), null, measurementList);
 					pathObjects.add(pathObject);
+					
+					roisCellsList.add(r);
 				}
 			} else {
 				pathObjects.addAll(nucleiObjects);
@@ -1016,8 +899,36 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 			for (PathObject pathObject : pathObjects)
 				pathObject.getMeasurementList().closeList();
 			
+			// Sometimes smoothing can cause nuclei of cell boundaries to be removed - in this case, 
+			// filter out the invalid ROIs now
+			int sizeBefore = pathObjects.size();
+			pathObjects.removeIf(p -> PathObjectTools.getROI(p, false).isEmpty() ||
+					PathObjectTools.getROI(p, true).isEmpty());
+			int sizeAfter = pathObjects.size();
+			if (sizeBefore != sizeAfter) {
+				logger.debug("Filtered out {} invalid cells (empty ROIs)", sizeBefore - sizeAfter);
+			}
+			
 			lastRunCompleted = true;
 		}
+		
+		
+		
+		
+		private static PolygonRoi smoothPolygonRoi(PolygonRoi r) {
+			FloatPolygon poly = r.getFloatPolygon();
+			FloatPolygon poly2 = new FloatPolygon();
+			int nPoints = poly.npoints;
+			for (int i = 0; i < nPoints; i += 2) {
+				int iMinus = (i + nPoints - 1) % nPoints;
+				int iPlus = (i + 1) % nPoints;
+				poly2.addPoint((poly.xpoints[iMinus] + poly.xpoints[iPlus] + poly.xpoints[i])/3, 
+						(poly.ypoints[iMinus] + poly.ypoints[iPlus] + poly.ypoints[i])/3);
+			}
+//			return new PolygonRoi(poly2, r.getType());
+			return new PolygonRoi(poly2, Roi.POLYGON);
+		}
+		
 		
 		
 		public List<PathObject> getPathObjects() {
@@ -1025,11 +936,16 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 		}
 		
 		
-		public void runDetection(double backgroundRadius, double maxBackground, double medianRadius, double sigma, double threshold, double minArea, double maxArea, boolean mergeAll, boolean watershedPostProcess, boolean excludeDAB, double cellExpansion, boolean limitExpansionByNucleusSize, boolean smoothBoundaries, boolean includeNuclei, boolean makeMeasurements) {
+//		public void runDetection(double backgroundRadius, double maxBackground, double medianRadius, double sigma, double threshold, double minArea, double maxArea, boolean mergeAll, boolean watershedPostProcess, boolean excludeDAB, double cellExpansion, boolean limitExpansionByNucleusSize, boolean smoothBoundaries, boolean includeNuclei, boolean makeMeasurements) {
+		public void runDetection(double backgroundRadius, double maxBackground, double medianRadius, double sigma, double threshold, double minArea, double maxArea, boolean mergeAll, boolean watershedPostProcess, boolean excludeDAB, double cellExpansion, boolean smoothBoundaries, boolean includeNuclei, boolean makeMeasurements, int z, int t) {
 			
 			boolean updateNucleusROIs = rois == null || bpLoG == null;
 			updateNucleusROIs = updateNucleusROIs ? updateNucleusROIs : this.medianRadius != medianRadius;
 			this.medianRadius = medianRadius;
+			
+			updateNucleusROIs = updateNucleusROIs ? updateNucleusROIs : this.t != t || this.z != z;
+			this.z = z;
+			this.t = t;
 			
 			updateNucleusROIs = updateNucleusROIs ? updateNucleusROIs : this.backgroundRadius != backgroundRadius;
 			this.backgroundRadius = backgroundRadius;
@@ -1072,29 +988,24 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 			updateAnything = updateAnything ? updateAnything : this.makeMeasurements != makeMeasurements;
 			this.makeMeasurements = makeMeasurements;
 			
-			updateAnything = updateAnything ? updateAnything : this.limitExpansionByNucleusSize != limitExpansionByNucleusSize;
-			this.limitExpansionByNucleusSize = limitExpansionByNucleusSize;
+//			updateAnything = updateAnything ? updateAnything : this.limitExpansionByNucleusSize != limitExpansionByNucleusSize;
+//			this.limitExpansionByNucleusSize = limitExpansionByNucleusSize;
 			
 //			if (!updateAnything)
 //				return;
 			
-			try {
-				doDetection(updateNucleusROIs);
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
+			doDetection(updateNucleusROIs);
+			
 		}
 		
 		
 	}
-
 	
 	
 	@Override
 	public String getDescription() {
-		return "Default cell detection algorithm for brightfield images with membrane staining";
+		return "Default cell detection algorithm for brightfield images with nuclear or cytoplasmic staining";
 	}
-
 
 
 	@Override
@@ -1115,10 +1026,9 @@ public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlu
 		if (Double.isNaN(pxSize))
 			return params.getDoubleParameterValue("cellExpansion") > 0 ? 25 : 10;
 		double cellExpansion = params.getDoubleParameterValue("cellExpansionMicrons") / pxSize;
-		int overlap = cellExpansion > 0 ? (int)(cellExpansion * 2) : 10;
+		int overlap = cellExpansion > 0 ? (int)(cellExpansion + 10) : 10;
 //		System.out.println("Tile overlap: " + overlap + " pixels");
 		return overlap;
 	}
-	
-	
+		
 }
