+++ qupath-core-processing/src/main/java/qupath/imagej/detect/cells/WatershedCellMembraneDetection.java
@@ -21,21 +21,20 @@
  * #L%
  */
 
-package qupath.imagej.detect.nuclei;
+package qupath.imagej.detect.cells;
 
 import java.awt.image.BufferedImage;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Iterator;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import ij.IJ;
 import ij.ImagePlus;
 import ij.Prefs;
 import ij.gui.PolygonRoi;
@@ -44,17 +43,17 @@ import ij.gui.Wand;
 import ij.measure.Calibration;
 import ij.measure.Measurements;
 import ij.plugin.filter.EDM;
+import ij.plugin.filter.MaximumFinder;
 import ij.plugin.filter.RankFilters;
 import ij.process.Blitter;
 import ij.process.ByteProcessor;
 import ij.process.ColorProcessor;
-import ij.process.FloatPolygon;
 import ij.process.FloatProcessor;
+import ij.process.FloodFiller;
 import ij.process.ImageProcessor;
 import ij.process.ImageStatistics;
 import ij.process.ShortProcessor;
 import qupath.imagej.color.ColorDeconvolutionIJ;
-import qupath.imagej.helpers.IJTools;
 import qupath.imagej.objects.PathImagePlus;
 import qupath.imagej.objects.ROIConverterIJ;
 import qupath.imagej.objects.measure.ObjectMeasurements;
@@ -68,7 +67,7 @@ import qupath.lib.analysis.algorithms.SimpleImage;
 import qupath.lib.analysis.stats.RunningStatistics;
 import qupath.lib.analysis.stats.StatisticsHelper;
 import qupath.lib.color.ColorDeconvolutionStains;
-import qupath.lib.color.StainVector;
+import qupath.lib.common.GeneralTools;
 import qupath.lib.images.ImageData;
 import qupath.lib.images.PathImage;
 import qupath.lib.images.servers.ImageServer;
@@ -77,11 +76,11 @@ import qupath.lib.measurements.MeasurementListFactory;
 import qupath.lib.measurements.MeasurementList;
 import qupath.lib.objects.PathObject;
 import qupath.lib.objects.PathObjects;
+import qupath.lib.objects.classes.PathClass;
 import qupath.lib.objects.classes.PathClassFactory;
 import qupath.lib.objects.helpers.PathObjectTools;
 import qupath.lib.plugins.AbstractTileableDetectionPlugin;
 import qupath.lib.plugins.ObjectDetector;
-import qupath.lib.plugins.parameters.DoubleParameter;
 import qupath.lib.plugins.parameters.Parameter;
 import qupath.lib.plugins.parameters.ParameterList;
 import qupath.lib.roi.PolygonROI;
@@ -90,16 +89,12 @@ import qupath.lib.roi.interfaces.PathArea;
 import qupath.lib.roi.interfaces.ROI;
 
 /**
- * Default command for cell detection within QuPath.
- * 
- * Assumes either a nuclear or cytoplasmic staining.
- * 
- * Quantification of membranous staining requires a separate detection command.
+ * Cell detection that takes into consideration membrane staining.
  * 
  * @author Pete Bankhead
  *
  */
-public class WatershedCellDetection extends AbstractTileableDetectionPlugin<BufferedImage> {
+public class WatershedCellMembraneDetection extends AbstractTileableDetectionPlugin<BufferedImage> {
 
 	private static String[] micronParameters = {
 		"requestedPixelSizeMicrons",
@@ -121,36 +116,25 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 		"cellExpansion",
 		};
 	
-	private static String[] fluorescenceParameters = {
-			"detectionImageFluorescence"
-	};
-
-	private static String[] brightfieldParameters = {
-			"detectionImageBrightfield",
-			"maxBackground"
-	};
-
 	transient private CellDetector detector;
 	
-	private final static Logger logger = LoggerFactory.getLogger(WatershedCellDetection.class);
-	
 	static String IMAGE_OPTICAL_DENSITY = "Optical density sum";
-	static String IMAGE_HEMATOXYLIN = "Hematoxylin OD";
+	static String IMAGE_HEMATOXYLIN = "Hematoxylin";
 	
-	ParameterList params;
+	private ParameterList params;
 	
 	
-	public WatershedCellDetection() {
+	public WatershedCellMembraneDetection() {
+		
+		Prefs.setThreads(1);
 		params = new ParameterList();
 		// TODO: Use a better way to determining if pixel size is available in microns
 //		params.addEmptyParameter("detectionParameters", "Detection parameters", true);
 
-		String microns = IJ.micronSymbol + "m";
+		String microns = GeneralTools.micrometerSymbol();
 		
-		params.addTitleParameter("Setup parameters");
-
-		params.addIntParameter("detectionImageFluorescence", "Choose detection channel", 1, null, "Choose the channel number containing a nucleus counterstain (e.g. DAPI)");
-
+		params.addEmptyParameter("paramsResolution", "Setup parameters", true);
+		
 		params.addChoiceParameter("detectionImageBrightfield", "Choose detection image", IMAGE_HEMATOXYLIN, Arrays.asList(IMAGE_HEMATOXYLIN, IMAGE_OPTICAL_DENSITY),
 				"Transformed image to which to apply the detection");
 
@@ -159,7 +143,7 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 //		params.addDoubleParameter("requestedPixelSize", "Requested downsample factor", 1, "");
 
 		
-		params.addTitleParameter("Nucleus parameters");
+		params.addEmptyParameter("paramsNuclei", "Nucleus parameters", true);
 		
 		params.addDoubleParameter("backgroundRadiusMicrons", "Background radius", 8, microns, 
 				"Radius for background estimation, should be > the largest nucleus radius, or <= 0 to turn off background subtraction");
@@ -169,7 +153,7 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 				"Sigma value for Gaussian filter used to reduce noise; increasing the value stops nuclei being fragmented, but may reduce the accuracy of boundaries");
 		params.addDoubleParameter("minAreaMicrons", "Minimum area", 10, microns+"^2",
 				"Detected nuclei with an area < minimum area will be discarded");
-		params.addDoubleParameter("maxAreaMicrons", "Maximum area", 400, microns+"^2",
+		params.addDoubleParameter("maxAreaMicrons", "Maximum area", 1000, microns+"^2",
 				"Detected nuclei with an area > maximum area will be discarded");
 
 		params.addDoubleParameter("backgroundRadius", "Background radius", 15, "px", 
@@ -183,36 +167,33 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 		params.addDoubleParameter("maxArea", "Maximum area", 1000, "px^2",
 				"Detected nuclei with an area > maximum area will be discarded");
 
-		params.addTitleParameter("Intensity parameters");
-		params.addDoubleParameter("threshold", "Threshold", 0.1, null,
+		params.addDoubleParameter("threshold", "Threshold", 0.1, null, 0, 2.5,
 				"Intensity threshold - detected nuclei must have a mean intensity >= threshold");
-//		params.addDoubleParameter("threshold", "Threshold", 0.1, null, 0, 2.5,
-//				"Intensity threshold - detected nuclei must have a mean intensity >= threshold");
 		params.addDoubleParameter("maxBackground", "Max background intensity", 2, null,
 				"If background radius > 0, detected nuclei occurring on a background > max background intensity will be discarded");
 		
 //		params.addBooleanParameter("mergeAll", "Merge all", true);
 		params.addBooleanParameter("watershedPostProcess", "Split by shape", true,
 				"Split merged detected nuclei based on shape ('roundness')");
-		params.addBooleanParameter("excludeDAB", "Exclude DAB (membrane staining)", false,
+		params.addBooleanParameter("excludeDAB", "Exclude DAB (membrane staining)", true,
 				"Set to 'true' if regions of high DAB staining should not be considered nuclei; useful if DAB stains cell membranes");
 		
 		
-		params.addTitleParameter("Cell parameters");
+		params.addEmptyParameter("paramsCells", "Cell parameters", true);
 
-		params.addDoubleParameter("cellExpansionMicrons", "Cell expansion", 5, microns, 0, 25,
+		params.addDoubleParameter("cellExpansionMicrons", "Cell expansion", 8, microns, 0, 25,
 				"Amount by which to expand detected nuclei to approximate the full cell area");
-		params.addDoubleParameter("cellExpansion", "Cell expansion", 5, "px",
+		params.addDoubleParameter("cellExpansion", "Cell expansion", 10, "px",
 				"Amount by which to expand detected nuclei to approximate the full cell area");
 		
-//		params.addBooleanParameter("limitExpansionByNucleusSize", "Limit cell expansion by nucleus size", false, "If checked, nuclei will not be expanded by more than their (estimated) smallest diameter in any direction - may give more realistic results for smaller, or 'thinner' nuclei");
+		params.addBooleanParameter("limitExpansionByNucleusSize", "Limit cell expansion by nucleus size", false, "If checked, nuclei will not be expanded by more than their (estimated) smallest diameter in any direction - may give more realistic results for smaller, or 'thinner' nuclei");
 			
 		params.addBooleanParameter("includeNuclei", "Include cell nucleus", true,
 				"If cell expansion is used, optionally include/exclude the nuclei within the detected cells");
 		
 		
-		params.addTitleParameter("General parameters");
-		params.addBooleanParameter("smoothBoundaries", "Smooth boundaries", true,
+		params.addEmptyParameter("paramsGeneral", "General parameters", true);
+		params.addBooleanParameter("smoothBoundaries", "Smooth boundaries", false,
 				"Smooth the detected nucleus/cell boundaries");
 		params.addBooleanParameter("makeMeasurements", "Make measurements", true,
 				"Add default shape & intensity measurements during detection");
@@ -221,21 +202,20 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 	
 	static class CellDetector implements ObjectDetector<BufferedImage> {
 	
-		private String lastServerPath = null;
-		//private PathImage<ImagePlus> pathImage; // Caching these cause out of memory errors...
-		private ROI pathROI;
-		
 		private List<PathObject> pathObjects = null;
-//		private WatershedCellDetector detector2;
-//		private FloatProcessor fpDetection, fpH, fpDAB;
-//		private ColorDeconvolutionStains stains;
 		
 		private boolean nucleiClassified = false;
 	
 			
 		public static double getPreferredPixelSizeMicrons(ImageData<BufferedImage> imageData, ParameterList params) {
-			if (imageData.getServer().hasPixelSizeMicrons())
-				return Math.max(params.getDoubleParameterValue("requestedPixelSizeMicrons"), imageData.getServer().getAveragedPixelSizeMicrons());
+			if (imageData.getServer().hasPixelSizeMicrons()) {
+				double requestedPixelSize = params.getDoubleParameterValue("requestedPixelSizeMicrons");
+				double averagedPixelSize = imageData.getServer().getAveragedPixelSizeMicrons();
+				if (requestedPixelSize < 0)
+					requestedPixelSize = averagedPixelSize * (-requestedPixelSize);
+				requestedPixelSize = Math.max(requestedPixelSize, averagedPixelSize);
+				return requestedPixelSize;
+			}
 			return Double.NaN;
 		}
 		
@@ -248,84 +228,49 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 			// Get a PathImage if we have a new ROI
 //			boolean imageChanged = false;
 			PathImage<ImagePlus> pathImage = null;
-			if (lastServerPath == null || !lastServerPath.equals(imageData.getServerPath()) || pathImage == null || !pathROI.equals(this.pathROI)) {
+//			if (pathImage == null || lastServerPath == null || !lastServerPath.equals(imageData.getServerPath()) || !pathROI.equals(this.pathROI)) {
 				ImageServer<BufferedImage> server = imageData.getServer();
-				lastServerPath = imageData.getServerPath();
-				pathImage = PathImagePlus.createPathImage(server, pathROI, ServerTools.getDownsampleFactor(server, getPreferredPixelSizeMicrons(imageData, params), true));
-				logger.trace("Cell detection with downsample: " + pathImage.getDownsampleFactor());
-				this.pathROI = pathROI;
+				pathImage = PathImagePlus.createPathImage(server, pathROI, ServerTools.getDownsampleFactor(server, getPreferredPixelSizeMicrons(imageData, params), false));
+//				System.out.println("Downsample: " + pathImage.getDownsampleFactor());
+//				this.pathROI = pathROI;
+//				lastServerPath = imageData.getServerPath();
 //				imageChanged = true;
-			}
+//			}
 			// Create a detector if we don't already have one for this image
 			boolean isBrightfield = imageData.isBrightfield();
-			//			if (detector2 == null || imageChanged || stains != imageData.getColorDeconvolutionStains()) {
-			//			if (imageChanged || stains != imageData.getColorDeconvolutionStains()) {
-			ImageProcessor ip = pathImage.getImage().getProcessor();
-			FloatProcessor fpDetection = null;
-			ColorDeconvolutionStains stains = imageData.getColorDeconvolutionStains();
-			Map<String, FloatProcessor> channels = new LinkedHashMap<>();
-			Map<String, FloatProcessor> channelsCell = new LinkedHashMap<>();
+			FloatProcessor fpDetection = null, fpH = null, fpDAB = null;
+//			if (detector2 == null || imageChanged || stains != imageData.getColorDeconvolutionStains()) {
+				ImageProcessor ip = pathImage.getImage().getProcessor();
+				ColorDeconvolutionStains stains = imageData.getColorDeconvolutionStains();
+				if (ip instanceof ColorProcessor && stains != null) {
+					
+					FloatProcessor[] fps = ColorDeconvolutionIJ.colorDeconvolve((ColorProcessor)ip, stains);
+					fpH = fps[0];
+					if (stains.isH_DAB())
+						fpDAB = fps[1];
+					else
+						fpDAB = null; // At this point, only DAB is quantified (eosin ignored for H&E)
+					
+					if (!params.getParameters().get("detectionImageBrightfield").isHidden()) {
+						if (params.getChoiceParameterValue("detectionImageBrightfield").equals(IMAGE_OPTICAL_DENSITY))
+							fpDetection = ColorDeconvolutionIJ.convertToOpticalDensitySum((ColorProcessor)ip, stains.getMaxRed(), stains.getMaxGreen(), stains.getMaxBlue());
+						else
+							fpDetection = (FloatProcessor)fpH.duplicate();
+					}
+
+					
+				}
+				if (fpDetection == null) {
+					// TODO: Deal with fluorescence
+					fpDetection = ip.convertToFloatProcessor();
+					fpH = ip.convertToFloatProcessor();
+					fpDAB = null;
+				}
+//			}
 			Roi roi = null;
 			if (pathROI != null)
 				roi = ROIConverterIJ.convertToIJRoi(pathROI, pathImage);
-			if (ip instanceof ColorProcessor && stains != null && isBrightfield) {
-				FloatProcessor[] fps = ColorDeconvolutionIJ.colorDeconvolve((ColorProcessor)ip, stains);
-				for (int i = 0; i < 3; i++) {
-					StainVector stain = stains.getStain(i+1);
-					if (!stain.isResidual()) {
-						channels.put(stain.getName() + " OD", fps[i]);
-						channelsCell.put(stain.getName() + " OD", fps[i]);
-					}
-				}
-//				channels.put("Hematoxylin OD",  fps[0]);
-//				if (stains.isH_DAB()) {
-//					channels.put("DAB OD", fps[1]);
-//					channelsCell.put("DAB OD", fps[1]);
-//				}
-//				else if (stains.isH_E()) {
-//					channels.put("Eosin OD", fps[1]);
-//					channelsCell.put("Eosin OD", fps[1]);
-//				}
-				
-
-				if (!params.getParameters().get("detectionImageBrightfield").isHidden()) {
-					if (params.getChoiceParameterValue("detectionImageBrightfield").equals(IMAGE_OPTICAL_DENSITY))
-						fpDetection = ColorDeconvolutionIJ.convertToOpticalDensitySum((ColorProcessor)ip, stains.getMaxRed(), stains.getMaxGreen(), stains.getMaxBlue());
-					else
-						fpDetection = (FloatProcessor)fps[0].duplicate();
-				}
-				
-				// Temporary test of the usefulness of RGB measurements...
-//				channels.put("Red", ((ColorProcessor)ip).toFloat(0, null));
-//				channels.put("Green", ((ColorProcessor)ip).toFloat(1, null));
-//				channels.put("Blue", ((ColorProcessor)ip).toFloat(2, null));
-				
-			} //else {
-			if (fpDetection == null) {
-				if (ip instanceof ColorProcessor) {
-					channels.put("Channel 1", ((ColorProcessor)ip).toFloat(0, null));
-					channels.put("Channel 2", ((ColorProcessor)ip).toFloat(1, null));
-					channels.put("Channel 3", ((ColorProcessor)ip).toFloat(2, null));
-				} else {
-					ImagePlus imp = pathImage.getImage();
-					for (int c = 1; c <= imp.getNChannels(); c++) {
-						channels.put("Channel " + c, imp.getStack().getProcessor(imp.getStackIndex(c, 0, 0)).convertToFloatProcessor());
-					}
-				}
-				// For fluorescence, measure everything
-				channelsCell.putAll(channels);
-				
-				// TODO: Deal with fluorescence... for now, defaults to first channel (may be totally wrong)
-				int detectionChannel = 1;
-				if (!isBrightfield)
-					detectionChannel = params.getIntParameterValue("detectionImageFluorescence");
-				fpDetection = channels.get("Channel " + detectionChannel);
-				if (fpDetection == null) {
-					logger.warn("Unable to find specified Channel {} - will default to Channel 1", detectionChannel);
-					fpDetection = channels.get("Channel 1");
-				}
-			}
-			WatershedCellDetector detector2 = new WatershedCellDetector(fpDetection, channels, channelsCell, roi, pathImage);
+			WatershedCellDetector detector2 = new WatershedCellDetector(fpDetection, fpH, fpDAB, roi, pathImage);
 			
 			// Create or reset the PathObjects list
 			if (pathObjects == null)
@@ -355,7 +300,7 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 			
 			detector2.runDetection(
 					backgroundRadius,
-					isBrightfield ? params.getDoubleParameterValue("maxBackground") : Double.NEGATIVE_INFINITY,
+					params.getDoubleParameterValue("maxBackground"),
 					medianRadius,
 					sigma,
 					params.getDoubleParameterValue("threshold"),
@@ -365,12 +310,10 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 					params.getBooleanParameterValue("watershedPostProcess"),
 					params.getBooleanParameterValue("excludeDAB"),
 					cellExpansion,
-//					params.getBooleanParameterValue("limitExpansionByNucleusSize"),
+					params.getBooleanParameterValue("limitExpansionByNucleusSize"),
 					params.getBooleanParameterValue("smoothBoundaries"),
 					params.getBooleanParameterValue("includeNuclei"),
-					params.getBooleanParameterValue("makeMeasurements"),
-					pathROI.getZ(),
-					pathROI.getT());// && isBrightfield);
+					params.getBooleanParameterValue("makeMeasurements") && isBrightfield);
 			
 			pathObjects.addAll(detector2.getPathObjects());
 					
@@ -410,26 +353,12 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 			map.get(name).setHidden(!pixelSizeKnown);
 		for (String name : pixelParameters)
 			map.get(name).setHidden(pixelSizeKnown);
-		
-		params.setHiddenParameters(!pixelSizeKnown, micronParameters);
-		params.setHiddenParameters(pixelSizeKnown, pixelParameters);
 
-		boolean isBrightfield = imageData.isBrightfield();
-		params.setHiddenParameters(!isBrightfield, brightfieldParameters);
-		params.setHiddenParameters(isBrightfield, fluorescenceParameters);
-		
-		if (!isBrightfield) {
-			if (imageData.getServer().getBitsPerPixel() > 8)
-				((DoubleParameter)params.getParameters().get("threshold")).setValue(100.0);
-			else
-				((DoubleParameter)params.getParameters().get("threshold")).setValue(25.0);
-		}
-
-//		map.get("detectionImageBrightfield").setHidden(imageData.getColorDeconvolutionStains() == null);
+		map.get("detectionImageBrightfield").setHidden(imageData.getColorDeconvolutionStains() == null);
 
 		map.get("excludeDAB").setHidden(imageData.getColorDeconvolutionStains() == null || !imageData.getColorDeconvolutionStains().isH_DAB());
 		
-//		map.get("makeMeasurements").setHidden(!imageData.isBrightfield());
+		map.get("makeMeasurements").setHidden(!imageData.isBrightfield());
 
 		return params;
 	}
@@ -453,14 +382,13 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 	
 	static class WatershedCellDetector {
 		
+		final private static Logger logger = LoggerFactory.getLogger(WatershedCellDetector.class);
 		
 		private boolean refineBoundary = true; // TODO: Consider making this variable accessible
 		
 		
 		private double backgroundRadius = 15;
 		private double maxBackground = 0.3;
-		
-		private int z = 0, t = 0;
 		
 		private boolean lastRunCompleted = false;
 		
@@ -478,15 +406,16 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 		private boolean excludeDAB = false;
 		private boolean smoothBoundaries = false;
 
-//		private boolean limitExpansionByNucleusSize = false;
+		private boolean limitExpansionByNucleusSize = false;
 
 		private boolean makeMeasurements = true;
 		
 		private Roi roi = null;
 		private FloatProcessor fpDetection = null;
-		private Map<String, FloatProcessor> channels = new LinkedHashMap<>(); // Map of channels to measure for nuclei only, and their names
-		private Map<String, FloatProcessor> channelsCell = new LinkedHashMap<>(); // Map of channels to measure for cell/cytoplasm, and their names
+		private FloatProcessor fpH = null;
+		private FloatProcessor fpDAB = null;
 		private ImageProcessor ipToMeasure = null;
+		private ImageProcessor ipBackground = null;
 		private List<PolygonRoi> rois = null;
 		private ByteProcessor bpLoG = null;
 		
@@ -495,15 +424,13 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 		
 		private PathImage<ImagePlus> pathImage = null;
 		
-		public WatershedCellDetector(FloatProcessor fpDetection, Map<String, FloatProcessor> channels, Map<String, FloatProcessor> channelsCell, Roi roi, PathImage<ImagePlus> pathImage) {
+		public WatershedCellDetector(FloatProcessor fpDetection, FloatProcessor fpH, FloatProcessor fpDAB, Roi roi, PathImage<ImagePlus> pathImage) {
 			this.fpDetection = fpDetection;
-			if (channels != null)
-				this.channels.putAll(channels);
-			if (channelsCell != null)
-				this.channelsCell.putAll(channelsCell);
+			this.fpH = fpH;
+			this.fpDAB = fpDAB;
 			this.roi = roi;
 			this.pathImage = pathImage;
-			Prefs.setThreads(1);
+//			Prefs.setThreads(1);
 		}
 		
 		
@@ -533,7 +460,10 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 						if (bpMask.getf(i) != 0f) {
 							ipBackground.setf(i, Float.NEGATIVE_INFINITY);
 						}
-					}
+					}				
+//				} else {
+//					// Don't return a mask - all pixels are ok
+//					System.out.println("Skipping background mask!");
 				}
 			}
 			
@@ -560,34 +490,63 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 				// Use Laplacian of Gaussian filtering followed by watershed transform to determine possible nucleus segments
 				// Result will be a dramatic over-segmentation...
 				FloatProcessor fpLoG = (FloatProcessor)fpDetection.duplicate();
+								
+////				fpLoG.copyBits(fpDAB, 0, 0, Blitter.ADD); // Testing is adding the optical densities helps...
+//				// Check we have some above-threshold pixels - if not, don't do more
+//				int indAbove = 0;
+//				int n = fpLoG.getWidth() * fpLoG.getHeight();
+//				while (indAbove < n && fpLoG.getf(indAbove) < threshold)
+//					indAbove++;
+//				if (indAbove == n)
+//					return;
 
 				// Start off with a median filter to reduce texture, if necessary
 				RankFilters rf = new RankFilters();
 				if (medianRadius > 0)
 					rf.rank(fpLoG, medianRadius, RankFilters.MEDIAN);
 
+				// Subtract background first, if needed
+				if (backgroundRadius > 0) {
+					
+//					// Could mask out definitely-background pixels if we wanted (and could be confident in background radius)... although it doesn't actually help all that much
+//					ByteProcessor bpPossible = SimpleThresholding.thresholdAboveEquals(fpH, (float)threshold);
+//					rf.rank(bpPossible, backgroundRadius, RankFilters.MAX);
+//					if (roi != null)
+//						ROILabeling.fillOutside(bpPossible, roi, 0);
+//					for (int i = 0; i < width * height; i++) {
+//						if (bpPossible.get(i) == 0)
+//							fpLoG.setf(i, Float.NEGATIVE_INFINITY);
+//					}
+////					bpPossible.max(1);
+////					new ImagePlus("Possible", bpPossible.duplicate()).show();
+
+//					ROILabeling.fillOutside(fpLoG, roi, Float.NEGATIVE_INFINITY);
+
+					ipBackground = fpLoG.duplicate();
+									
+					bpBackgroundMask = limitedOpeningByReconstruction(fpLoG, ipBackground, backgroundRadius, maxBackground);
+					
+//					ipBackground = MorphologicalReconstructionInteger.openingByReconstruction(fpLoG, backgroundRadius);
+					fpLoG.copyBits(ipBackground, 0, 0, Blitter.SUBTRACT);
+					ipToMeasure = fpLoG.duplicate();
+//					new ImagePlus("Background", ipBackground.duplicate()).show();
+				} else {
+					ipToMeasure = fpDetection;
+					ipBackground = null;
+				}
+				
 				//--------NEW--------
-				if (excludeDAB && channels.containsKey("Hematoxylin OD") && channels.containsKey("DAB OD")) {
+				if (excludeDAB && fpH != null && fpDAB != null) {
 					// If we are avoiding DAB, set pixels away from potential nuclei to zero
-					FloatProcessor fpDAB = channels.get("DAB OD");
 					fpDAB.setRoi(roi);
-					ByteProcessor bpH = SimpleThresholding.greaterThanOrEqual(channels.get("Hematoxylin OD"), fpDAB);
+					ByteProcessor bpH = SimpleThresholding.greaterThanOrEqual(fpH, fpDAB);
 					bpH.multiply(1.0/255.0);
-					rf.rank(bpH, 2.5, RankFilters.MEDIAN);
+					rf.rank(bpH, 2.5, RankFilters.MEDIAN); // TODO: Check hard-coded filter sizes for reasonableness
 					rf.rank(bpH, 2.5, RankFilters.MAX);
 					fpLoG.copyBits(bpH, 0, 0, Blitter.MULTIPLY);
 				}
+//				new ImagePlus("Log", fpLoG.duplicate()).show();
 				//--------END_NEW--------
-				
-				// Subtract background first, if needed
-				if (backgroundRadius > 0) {
-					ImageProcessor ipBackground = fpLoG.duplicate();
-					bpBackgroundMask = limitedOpeningByReconstruction(fpLoG, ipBackground, backgroundRadius, maxBackground);
-					fpLoG.copyBits(ipBackground, 0, 0, Blitter.SUBTRACT);
-					ipToMeasure = fpLoG.duplicate();
-				} else {
-					ipToMeasure = fpDetection;
-				}
 				
 				// Apply (approximation of) Laplacian of Gaussian filter
 				fpLoG.blurGaussian(sigma);
@@ -596,31 +555,30 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 				// Threshold the main LoG image
 				bpLoG = SimpleThresholding.thresholdAbove(fpLoG, 0f);
 				// Need to set the threshold very slightly above zero for ImageJ
-				// TODO: DECIDE ON USING MY WATERSHED OR IMAGEJ'S....
 				fpLoG.setRoi(roi);
+				
+//				ROILabeling.fillOutside(fpLoG, roi, Float.NEGATIVE_INFINITY);
 				
 				ImageProcessor ipTemp = RegionalExtrema.findRegionalMaxima(fpLoG, 0.001f, false);
 				ImageProcessor ipLabels = ROILabeling.labelImage(ipTemp, 0, false);
+//				new ImagePlus("Labels before", ipLabels.duplicate()).show();
 				Watershed.doWatershed(fpLoG, ipLabels, 0, false);
+//				new ImagePlus("Labels after", ipLabels.duplicate()).show();
+				
 				
 				ipLabels.setThreshold(0.5, Double.POSITIVE_INFINITY, ImageProcessor.NO_LUT_UPDATE);
 				// TODO: Consider 4/8 connectivity for watershed nucleus ROIs
 				rois = ROILabeling.getFilledPolygonROIs(ipLabels, Wand.FOUR_CONNECTED);			
+//				rois = ROILabeling.getFilledPolygonROIsExperimental(ipLabels);
 				
+				
+//				new ImagePlus("Labels", ipLabels.duplicate()).show();
 				if (Thread.currentThread().isInterrupted())
 					return;
 			} 
 			
 			if (bp == null)
-				bp = new ByteProcessor(width, height);	
-			
-//			// TODO: Consider application of an automated threshold
-//			if (threshold < 0) {
-//				ipToMeasure.resetRoi();
-//				ImageStatistics stats = ipToMeasure.getStatistics();
-//				threshold = stats.mean;// + stats.stdDev;
-//				logger.info("Mean threshold set: " + threshold);
-//			}
+				bp = new ByteProcessor(width, height);			
 
 			bp.setValue(255);
 			for (Roi r : rois) {
@@ -636,9 +594,16 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 					if (bpBackgroundMask.getStatistics().mean > 0)
 						continue;				
 				}
+//				if (ipBackground != null && !Double.isNaN(maxBackground)) {
+//					ipBackground.setRoi(r);
+//					if (ipBackground.getStatistics().mean > maxBackground)
+//						continue;
+//				}
 				// Fill the ROI to keep it
 				bp.fill(r);
 			}
+			
+//			new ImagePlus("BP early", bp.duplicate()).show();
 			
 			if (Thread.currentThread().isInterrupted())
 				return;
@@ -660,31 +625,34 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 			// TODO: Look at the better boundary clearing implemented in Fast_nucleus_counts
 			if (roi != null)
 				ROILabeling.clearOutside(bp, roi);
+//				ROILabeling.clearBoundary(bp, roi, 0);
 			
 			// Locate nucleus ROIs
 			bp.setThreshold(127, Double.POSITIVE_INFINITY, ImageProcessor.NO_LUT_UPDATE);
 			
 			
-			if (IJ.debugMode) {
-				IJTools.quickShowImage("Binary", bp.duplicate());
-			}
+//			new ImagePlus("BP later", bp.duplicate()).show();
 
 			//----------------------------
 			// MINOR BOUNDARY REFINEMENT
 			// The idea is that Gaussian smoothing tends to cause the boundaries of 'thin' nuclei to be overestimated;
 			// this uses a smaller filter to correct instances where the boundary has moved by just one pixel
 			if (refineBoundary && sigma > 1.5) {
-				FloatProcessor fpBoundaryCleanup = (FloatProcessor)fpDetection.duplicate();
-				fpBoundaryCleanup.blurGaussian(1);
-				fpBoundaryCleanup.convolve(new float[]{0, -1, 0, -1, 4, -1, 0, -1, 0}, 3, 3);
-				ByteProcessor bp2 = SimpleThresholding.thresholdAbove(fpBoundaryCleanup, 0f);
+				
+//				new ImagePlus("Before", bp.duplicate()).show();
+				
+				FloatProcessor fpLoG = (FloatProcessor)fpDetection.duplicate();
+				fpLoG.blurGaussian(1);
+				fpLoG.convolve(new float[]{0, -1, 0, -1, 4, -1, 0, -1, 0}, 3, 3);
+				ByteProcessor bp2 = SimpleThresholding.thresholdAbove(fpLoG, 0f);
 				bp2.copyBits(bp, 0, 0, Blitter.MIN); // Remove everything not detected in bp
 				bp.filter(ByteProcessor.MIN);
 				bp.copyBits(bp2, 0, 0, Blitter.MAX);
+
+//				new ImagePlus("After", bp.duplicate()).show();
+
 				regenerateROIs = true;
 			}
-			
-			//----------------------------
 			
 			roisNuclei = ROILabeling.getFilledPolygonROIs(bp, Wand.FOUR_CONNECTED);
 
@@ -697,15 +665,14 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 				Iterator<PolygonRoi> iter = roisNuclei.iterator();
 				while (iter.hasNext()) {
 					Roi roiTemp = iter.next();
-					ipToMeasure.setRoi(roiTemp);
-					ImageStatistics stats = ImageStatistics.getStatistics(ipToMeasure, Measurements.AREA | Measurements.MEAN, null);
-					double area = stats.pixelCount;
-					if ((stats.mean < threshold) || (minArea > 0 && area < minArea) || (maxArea > 0 && area > maxArea)) {
+					fpDetection.setRoi(roiTemp);
+					double area = ImageStatistics.getStatistics(fpDetection, Measurements.AREA, null).pixelCount;
+					if ((minArea > 0 && area < minArea) || (maxArea > 0 && area > maxArea)) {
 						iter.remove();
 						bp.fill(roiTemp);
 					}
 				}
-				ipToMeasure.resetRoi();
+				fpDetection.resetRoi();
 			}
 			
 			
@@ -713,14 +680,98 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 			ShortProcessor ipLabels = new ShortProcessor(width, height);
 			ROILabeling.labelROIs(ipLabels, roisNuclei);
 			
-			// Measure nuclei for all required channels
-			Map<String, List<RunningStatistics>> statsMap = new LinkedHashMap<>();
+			
+			
+			
+			//----------------------------
+			Map<Float, PolygonRoi> roisCellsList = null;
+			ImageProcessor ipLabelsCells = null;
+			int nCells = 0;
+			
+			// Membrane detection (new 9/8/2015)
+			if (excludeDAB && fpDAB != null) {
+				
+				FloatProcessor fpMembranes = (FloatProcessor)fpDAB.duplicate();
+				fpMembranes.blurGaussian(2); // TODO: Enable a different sigma parameter?
+				
+				float membraneThreshold = .2f; // TODO: Enable a different membrane threshold?
+				for (int i = 0; i < width * height; i++) {
+					fpMembranes.setf(i, membraneThreshold - fpMembranes.getf(i));
+				}
+				fpMembranes.max(0.);
+				
+				// Fill nuclei with max values
+				fpMembranes.setValue(0.1);
+				for (Roi r : roisNuclei) {
+					fpMembranes.fill(r);
+				}
+				
+				// Locate regional maxima that will be used for detection
+				ByteProcessor bpMarkers = new MaximumFinder().findMaxima(fpMembranes, 0.09, ImageProcessor.NO_THRESHOLD, MaximumFinder.IN_TOLERANCE, false, false);
+				
+				// Remove any impossibly-large areas
+				ROILabeling.removeByAreas(bpMarkers, 1, maxArea, true);
+				
+				// Determine a mask of the furthest anything can expand
+				ByteProcessor bpMaxExpansion = (ByteProcessor)bpMarkers.duplicate();
+				new RankFilters().rank(bpMaxExpansion, cellExpansion, RankFilters.MAX);
+				
+				// Update the mask to remove completely unstained pixels
+				for (int i = 0; i < width * height; i++) {
+					if (fpH.getf(i) + fpDAB.getf(i) < 0.025f)
+						bpMaxExpansion.set(i, 0);
+				}
+				
+				// Expand all potential cells as much as possible
+				float minThreshold = (float)(fpMembranes.getStatistics().min - 0.05);
+				for (int i = 0; i < width * height; i++) {
+					if (bpMaxExpansion.get(i) == 0)
+						fpMembranes.setf(i, minThreshold);
+				}
+				fpMembranes.resetMinAndMax();
+				int lastLabel = roisNuclei.size();
+				ipLabelsCells = ipLabels.duplicate();
+				for (int i = 0; i < width * height; i++) {
+					if (bpMarkers.get(i) != 0 && ipLabelsCells.get(i) == 0) {
+						ipLabelsCells.set(i, Short.MAX_VALUE);
+					}
+				}
+				FloodFiller ff = new FloodFiller(ipLabelsCells);
+				for (int i = 0; i < width * height; i++) {
+					if (ipLabelsCells.get(i) == Short.MAX_VALUE) {
+						lastLabel++;
+						ipLabelsCells.setValue(lastLabel);
+						ff.fill(i%width, i/width);
+					}
+				}
+				
+//				fpMembranes.resetMinAndMax();
+//				new ImagePlus("BEFORE", fpMembranes.duplicate()).show();
+//				ipLabels = ROILabeling.labelImage(bpMarkers, false);
+				Watershed.doWatershed(fpMembranes, ipLabelsCells, minThreshold+.025, false);
+				ipLabelsCells.setThreshold(0.5, Double.POSITIVE_INFINITY, ImageProcessor.NO_LUT_UPDATE);
+				roisCellsList = ROILabeling.getFilledPolygonROIsFromLabels(ipLabelsCells, Wand.FOUR_CONNECTED);
+				nCells = lastLabel;
+			}
+			
+			
+			
+			//----------------------------
+			
+			
+			
+			
+			
+			// Measure nuclei
+			List<RunningStatistics> statsHematoxylin = null;
+			List<RunningStatistics> statsDAB = null;
 			if (makeMeasurements) {
+				statsHematoxylin = StatisticsHelper.createRunningStatisticsList(roisNuclei.size());
 				SimpleImage imgLabels = new PixelImageIJ(ipLabels);
-				for (String key : channels.keySet()) {
-					List<RunningStatistics> statsList = StatisticsHelper.createRunningStatisticsList(roisNuclei.size());
-					StatisticsHelper.computeRunningStatistics(new PixelImageIJ(channels.get(key)), imgLabels, statsList);
-					statsMap.put(key, statsList);
+				StatisticsHelper.computeRunningStatistics(new PixelImageIJ(fpH), imgLabels, statsHematoxylin);
+				if (fpDAB != null) {
+					statsDAB = StatisticsHelper.createRunningStatisticsList(roisNuclei.size());
+					StatisticsHelper.computeRunningStatistics(new PixelImageIJ(fpDAB), imgLabels, statsDAB);
 				}
 			}
 			
@@ -733,37 +784,49 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 			List<PathObject> nucleiObjects = new ArrayList<>();
 			Calibration cal = pathImage.getImage().getCalibration();
 			for (int i = 0; i < roisNuclei.size(); i++) {
-				PolygonRoi rOrig = roisNuclei.get(i);
+				PolygonRoi r = roisNuclei.get(i);
 				
-				PolygonRoi r = rOrig;
 				if (smoothBoundaries)
-					r = new PolygonRoi(rOrig.getInterpolatedPolygon(Math.min(2.5, rOrig.getNCoordinates()*0.1), true), Roi.POLYGON);
+					r = new PolygonRoi(r.getInterpolatedPolygon(Math.min(2.5, r.getNCoordinates()*0.1), true), Roi.POLYGON);
 				
-				PolygonROI pathROI = ROIConverterIJ.convertToPolygonROI(r, cal, pathImage.getDownsampleFactor(), 0, z, t);
+				PolygonROI pathROI = ROIConverterIJ.convertToPolygonROI(r, cal, pathImage.getDownsampleFactor());
 				
 				if (smoothBoundaries) {
+//					int nBefore = pathROI.nVertices();
 					pathROI = ShapeSimplifier.simplifyPolygon(pathROI, pathImage.getDownsampleFactor()/4.0);
+//					int nAfter = pathROI.nVertices();
+//					System.out.println("Vertices removed: " + (nBefore - nAfter));
 				}
 				
 				// Create a new shared measurement list
 				MeasurementList measurementList = MeasurementListFactory.createMeasurementList(makeMeasurements ? 30 : 0, MeasurementList.TYPE.FLOAT);
 				
 				if (makeMeasurements) {
-					ObjectMeasurements.addShapeStatistics(measurementList, r, fpDetection, cal, "Nucleus: ");
+					ObjectMeasurements.addShapeStatistics(measurementList, r, fpH, cal, "Nucleus: ");
 	
-					for (String key : channels.keySet()) {
-						List<RunningStatistics> statsList = statsMap.get(key);
-						RunningStatistics stats = statsList.get(i);
-						measurementList.addMeasurement("Nucleus: " + key + " mean", stats.getMean());
-						measurementList.addMeasurement("Nucleus: " + key + " sum", stats.getSum());
-						measurementList.addMeasurement("Nucleus: " + key + " std dev", stats.getStdDev());
-						measurementList.addMeasurement("Nucleus: " + key + " max", stats.getMax());
-						measurementList.addMeasurement("Nucleus: " + key + " min", stats.getMin());
-						measurementList.addMeasurement("Nucleus: " + key + " range", stats.getRange());
+	//				PathObject pathObject = new PathDetectionObject(pathROI, PathPrefs.getNucleusClass());
+	//				PathObjectIJ.computeShapeStatistics(pathObject, pathImage, fpH, pathImage.getImage().getCalibration(), "Nucleus: ");
+					
+					RunningStatistics stats = statsHematoxylin.get(i);
+					measurementList.addMeasurement("Nucleus: Hematoxylin OD mean", stats.getMean());
+					measurementList.addMeasurement("Nucleus: Hematoxylin OD sum", stats.getSum());
+					measurementList.addMeasurement("Nucleus: Hematoxylin OD std dev", stats.getStdDev());
+					measurementList.addMeasurement("Nucleus: Hematoxylin OD max", stats.getMax());
+					measurementList.addMeasurement("Nucleus: Hematoxylin OD min", stats.getMin());
+					measurementList.addMeasurement("Nucleus: Hematoxylin OD range", stats.getRange());
+					if (statsDAB != null) {
+						stats = statsDAB.get(i);
+						measurementList.addMeasurement("Nucleus: DAB OD mean", stats.getMean());
+						measurementList.addMeasurement("Nucleus: DAB OD sum", stats.getSum());
+						measurementList.addMeasurement("Nucleus: DAB OD std dev", stats.getStdDev());
+						measurementList.addMeasurement("Nucleus: DAB OD max", stats.getMax());
+						measurementList.addMeasurement("Nucleus: DAB OD min", stats.getMin());
+						measurementList.addMeasurement("Nucleus: DAB OD range", stats.getRange());
 					}
 				}
 				
 				// TODO: It would be more efficient to measure the hematoxylin intensities along with the shapes
+//				PathObject pathObject = new PathDetectionObject(pathROI, PathClassFactory.getNucleusClass(), measurementList);
 				PathObject pathObject = PathObjects.createDetectionObject(pathROI, null, measurementList);
 				nucleiObjects.add(pathObject);
 
@@ -772,101 +835,169 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 			if (Thread.currentThread().isInterrupted())
 				return;
 
-			List<Roi> roisCellsList = null;
+//			List<Roi> roisCellsList = null;
 			
 			// Optionally expand the nuclei to become cells
 			if (cellExpansion > 0) {
-				FloatProcessor fpEDM = new EDM().makeFloatEDM(bp, (byte)255, false);
-				fpEDM.multiply(-1);
 				
-				double cellExpansionThreshold = -cellExpansion;
+				Map<Float, PolygonRoi> roisCells;
+				if (roisCellsList == null) {
 				
-				// Create cell ROIs
-				ImageProcessor ipLabelsCells = ipLabels.duplicate();
-				Watershed.doWatershed(fpEDM, ipLabelsCells, cellExpansionThreshold, false);
-				PolygonRoi[] roisCells = ROILabeling.labelsToFilledROIs(ipLabelsCells, roisNuclei.size());
+					FloatProcessor fpEDM = new EDM().makeFloatEDM(bp, (byte)255, false);
+					fpEDM.multiply(-1);
+					
+					// ---------------------- EXPERIMENTAL - CONSTRAIN CELL EXPANSION BY NUCLEUS SIZE
+					double cellExpansionThreshold = -cellExpansion;
+					if (limitExpansionByNucleusSize) {
+						MaximumFinder mf = new MaximumFinder();
+						ByteProcessor bpVoronoi = mf.findMaxima(fpEDM, .5, cellExpansionThreshold, MaximumFinder.SEGMENTED, false, false);
+						FloatProcessor fpEDM2 = new EDM().makeFloatEDM(bp, (byte)0, false);
+						
+						FloatProcessor fpMarkers = new FloatProcessor(fpEDM2.getWidth(), fpEDM2.getHeight());
+						Arrays.fill((float[])(fpMarkers.getPixels()), Float.NEGATIVE_INFINITY);
+						for (Roi r : roisNuclei) {
+							fpEDM2.setRoi(r);
+							double max = fpEDM2.getStatistics().max;
+							fpMarkers.setValue(-2. * max);
+							fpMarkers.fill(r);
+						}
+						
+						for (int i = 0; i < bpVoronoi.getWidth()*bpVoronoi.getHeight(); i++) {
+							if (bpVoronoi.getf(i) == 0)
+								fpEDM.setf(i, Float.NEGATIVE_INFINITY);
+						}
+						if (!MorphologicalReconstruction.morphologicalReconstruction(fpMarkers, fpEDM))
+							logger.error("Problem during morphological reconstruction!");
+						ByteProcessor bpMax = mf.findMaxima(fpMarkers, 0.0001, MaximumFinder.IN_TOLERANCE, false);
+						fpEDM.copyBits(bpMax, 0, 0, Blitter.COPY);
+						cellExpansionThreshold = 128; // Cell expansion has already been applied
+					}
+					
+					// ---------------------- END EXPERIMENTAL
+					
+					// Create cell ROIs
+					ipLabelsCells = ipLabels.duplicate();
+					Watershed.doWatershed(fpEDM, ipLabelsCells, cellExpansionThreshold, false);
+					roisCells = ROILabeling.getFilledPolygonROIsFromLabels(ipLabelsCells, roisNuclei.size());
+					nCells = roisCells.size();
+				} else {
+					roisCells = roisCellsList;
+				}
+				
+				
+				
 				
 				// Compute cell DAB stats
-				Map<String, List<RunningStatistics>> statsMapCell = new LinkedHashMap<>();
-				if (makeMeasurements) {
-					for (String key : channelsCell.keySet()) {
-						List<RunningStatistics> statsList = StatisticsHelper.createRunningStatisticsList(roisNuclei.size());
-						StatisticsHelper.computeRunningStatistics(new PixelImageIJ(channelsCell.get(key)), new PixelImageIJ(ipLabelsCells), statsList);
-						statsMapCell.put(key, statsList);
-					}
+				List<RunningStatistics> statsDABCell = null;
+				if (fpDAB != null && makeMeasurements) {
+					statsDABCell = StatisticsHelper.createRunningStatisticsList(nCells);
+					StatisticsHelper.computeRunningStatistics(new PixelImageIJ(fpDAB), new PixelImageIJ(ipLabelsCells), statsDABCell);
 				}
 							
 				// Create labelled image for cytoplasm, i.e. remove all nucleus pixels
-				// TODO: Make a buffer zone between nucleus and cytoplasm!
+				// TODO: Make a buffer zone between nucleus and cytoplasm?!
 				for (int i = 0; i < ipLabels.getWidth() * ipLabels.getHeight(); i++) {
 					if (ipLabels.getf(i) != 0)
 						ipLabelsCells.setf(i, 0f);
 				}
 				
 				// Compute cytoplasm stats
-				Map<String, List<RunningStatistics>> statsMapCytoplasm = new LinkedHashMap<>();
-				if (makeMeasurements) {
-					for (String key : channelsCell.keySet()) {
-						List<RunningStatistics> statsList = StatisticsHelper.createRunningStatisticsList(roisNuclei.size());
-						StatisticsHelper.computeRunningStatistics(new PixelImageIJ(channelsCell.get(key)), new PixelImageIJ(ipLabelsCells), statsList);
-						statsMapCytoplasm.put(key, statsList);
+				List<RunningStatistics> statsDABCytoplasm = null;
+				if (includeNuclei && fpDAB != null && makeMeasurements) {
+					statsDABCytoplasm = StatisticsHelper.createRunningStatisticsList(nCells);
+					StatisticsHelper.computeRunningStatistics(new PixelImageIJ(fpDAB), new PixelImageIJ(ipLabelsCells), statsDABCytoplasm);
+				}
+				
+				// Create membrane stats
+				List<RunningStatistics> statsDABMembrane = null;
+				if (includeNuclei && excludeDAB && fpDAB != null && makeMeasurements) {
+					statsDABMembrane = StatisticsHelper.createRunningStatisticsList(nCells);
+					ImageProcessor ipLabelsMembrane = new ShortProcessor(width, height);
+					// TODO: WARNING!  This method of creating measurements doesn't permit membranes to overlap.
+					// This means that sometimes one cell could 'steal' a bit of the membrane of another cell.
+					// However it's unlikely this actually make any substantial difference...
+					for (Entry<Float, PolygonRoi> entry : roisCells.entrySet()) {
+						Float label = entry.getKey();
+						PolygonRoi roiTemp = entry.getValue();
+						ipLabelsMembrane.setValue(label.doubleValue());
+						ipLabelsMembrane.draw(roiTemp);
 					}
+					StatisticsHelper.computeRunningStatistics(new PixelImageIJ(fpDAB), new PixelImageIJ(ipLabelsMembrane), statsDABMembrane);
+//					ipLabelsMembrane.resetMinAndMax();
+//					new ImagePlus("LABELS", ipLabelsMembrane.duplicate()).show();
 				}
 				
 				
 				// Create cell objects
-				roisCellsList = new ArrayList<>(roisCells.length); // In case we need texture measurements, store all cell ROIs
-				for (int i = 0; i < roisCells.length; i++) {
-					PolygonRoi r = roisCells[i];
+//				PathClass cellClass = PathClassFactory.getPathClass("Cell", ColorTools.makeRGB(255, 200, 0));
+				PathClass cellClass = null;
+				for (Entry<Float, PolygonRoi> entry : roisCells.entrySet()) {
+					PolygonRoi r = entry.getValue();
 					if (r == null)
 						continue;
 					if (smoothBoundaries)
 						r = new PolygonRoi(r.getInterpolatedPolygon(Math.min(2.5, r.getNCoordinates()*0.1), false), Roi.POLYGON); // TODO: Check this smoothing - it can be troublesome, causing nuclei to be outside cells
 //						r = smoothPolygonRoi(r);
 
-					PolygonROI pathROI = ROIConverterIJ.convertToPolygonROI(r, pathImage.getImage().getCalibration(), pathImage.getDownsampleFactor(), 0, z, t);
+					PolygonROI pathROI = ROIConverterIJ.convertToPolygonROI(r, pathImage.getImage().getCalibration(), pathImage.getDownsampleFactor());
 					if (smoothBoundaries)
 						pathROI = ShapeSimplifier.simplifyPolygon(pathROI, pathImage.getDownsampleFactor()/4.0);
 
 					
 					MeasurementList measurementList = null;
 					PathObject nucleus = null;
-					if (includeNuclei) {
+					int label = entry.getKey().intValue();
+					if (label < nucleiObjects.size())
+						nucleus = nucleiObjects.get(label-1);
+
+					// If we don't have a nucleus, check the cell area isn't too large
+					if (nucleus == null && pathROI.getArea() > maxArea * 2)
+						continue;
+
+					// Prepare measurement list with/without nucleus
+					if (includeNuclei && nucleus != null) {
 						// Use the nucleus' measurement list
-						nucleus = nucleiObjects.get(i);
 						measurementList = nucleus.getMeasurementList();					
 					} else {
 						// Create a new measurement list
 						measurementList = MeasurementListFactory.createMeasurementList(makeMeasurements ? 12 : 0, MeasurementList.TYPE.GENERAL);
+						nucleus = null;
 					}
-									
+					
+					
 					// Add cell shape measurements
 					if (makeMeasurements) {
 						ObjectMeasurements.addShapeStatistics(measurementList, r, fpDetection, pathImage.getImage().getCalibration(), "Cell: ");
 	//					ObjectMeasurements.computeShapeStatistics(pathObject, pathImage, fpH, pathImage.getImage().getCalibration());
 	
 						// Add cell measurements
-						for (String key : channelsCell.keySet()) {
-							if (statsMapCell.containsKey(key)) {
-								RunningStatistics stats = statsMapCell.get(key).get(i);
-								measurementList.addMeasurement("Cell: " + key + " mean", stats.getMean());
-								measurementList.addMeasurement("Cell: " + key + " std dev", stats.getStdDev());
-								measurementList.addMeasurement("Cell: " + key + " max", stats.getMax());
-								measurementList.addMeasurement("Cell: " + key + " min", stats.getMin());
-		//						pathObject.addMeasurement("Cytoplasm: " + key + " range", stats.getRange());
-							}
+						if (statsDABCell != null) {
+							RunningStatistics stats = statsDABCell.get(label-1);
+							measurementList.addMeasurement("Cell: DAB OD mean", stats.getMean());
+							measurementList.addMeasurement("Cell: DAB OD std dev", stats.getStdDev());
+							measurementList.addMeasurement("Cell: DAB OD max", stats.getMax());
+							measurementList.addMeasurement("Cell: DAB OD min", stats.getMin());
+	//						pathObject.addMeasurement("Cytoplasm: DAB OD range", stats.getRange());
 						}
-							
-							// Add cytoplasm measurements
-						for (String key : channelsCell.keySet()) {
-							if (statsMapCytoplasm.containsKey(key)) {
-								RunningStatistics stats = statsMapCytoplasm.get(key).get(i);
-								measurementList.addMeasurement("Cytoplasm: " + key + " mean", stats.getMean());
-								measurementList.addMeasurement("Cytoplasm: " + key + " std dev", stats.getStdDev());
-								measurementList.addMeasurement("Cytoplasm: " + key + " max", stats.getMax());
-								measurementList.addMeasurement("Cytoplasm: " + key + " min", stats.getMin());
-		//						pathObject.addMeasurement("Cytoplasm: " + key + " range", stats.getRange());
-							}
+						
+						// Add cytoplasm measurements
+						if (statsDABCytoplasm != null) {
+							RunningStatistics stats = statsDABCytoplasm.get(label-1);
+							measurementList.addMeasurement("Cytoplasm: DAB OD mean", stats.getMean());
+							measurementList.addMeasurement("Cytoplasm: DAB OD std dev", stats.getStdDev());
+							measurementList.addMeasurement("Cytoplasm: DAB OD max", stats.getMax());
+							measurementList.addMeasurement("Cytoplasm: DAB OD min", stats.getMin());
+	//						pathObject.addMeasurement("Cytoplasm: DAB OD range", stats.getRange());
+						}
+						
+						// Add membrane measurements
+						if (statsDABMembrane != null) {
+							RunningStatistics stats = statsDABMembrane.get(label-1);
+							measurementList.addMeasurement("Membrane: DAB OD mean", stats.getMean());
+							measurementList.addMeasurement("Membrane: DAB OD std dev", stats.getStdDev());
+							measurementList.addMeasurement("Membrane: DAB OD max", stats.getMax());
+							measurementList.addMeasurement("Membrane: DAB OD min", stats.getMin());
+	//						pathObject.addMeasurement("Cytoplasm: DAB OD range", stats.getRange());
 						}
 						
 						// Add nucleus area ratio, if available
@@ -880,10 +1011,8 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 
 					
 					// Create & store the cell object
-					PathObject pathObject = PathObjects.createCellObject(pathROI, nucleus == null ? null : nucleus.getROI(), null, measurementList);
+					PathObject pathObject = PathObjects.createCellObject(pathROI, nucleus == null ? null : nucleus.getROI(), cellClass, measurementList);
 					pathObjects.add(pathObject);
-					
-					roisCellsList.add(r);
 				}
 			} else {
 				pathObjects.addAll(nucleiObjects);
@@ -893,36 +1022,8 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 			for (PathObject pathObject : pathObjects)
 				pathObject.getMeasurementList().closeList();
 			
-			// Sometimes smoothing can cause nuclei of cell boundaries to be removed - in this case, 
-			// filter out the invalid ROIs now
-			int sizeBefore = pathObjects.size();
-			pathObjects.removeIf(p -> PathObjectTools.getROI(p, false).isEmpty() ||
-					PathObjectTools.getROI(p, true).isEmpty());
-			int sizeAfter = pathObjects.size();
-			if (sizeBefore != sizeAfter) {
-				logger.debug("Filtered out {} invalid cells (empty ROIs)", sizeBefore - sizeAfter);
-			}
-			
 			lastRunCompleted = true;
 		}
-		
-		
-		
-		
-		private static PolygonRoi smoothPolygonRoi(PolygonRoi r) {
-			FloatPolygon poly = r.getFloatPolygon();
-			FloatPolygon poly2 = new FloatPolygon();
-			int nPoints = poly.npoints;
-			for (int i = 0; i < nPoints; i += 2) {
-				int iMinus = (i + nPoints - 1) % nPoints;
-				int iPlus = (i + 1) % nPoints;
-				poly2.addPoint((poly.xpoints[iMinus] + poly.xpoints[iPlus] + poly.xpoints[i])/3, 
-						(poly.ypoints[iMinus] + poly.ypoints[iPlus] + poly.ypoints[i])/3);
-			}
-//			return new PolygonRoi(poly2, r.getType());
-			return new PolygonRoi(poly2, Roi.POLYGON);
-		}
-		
 		
 		
 		public List<PathObject> getPathObjects() {
@@ -930,16 +1031,11 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 		}
 		
 		
-//		public void runDetection(double backgroundRadius, double maxBackground, double medianRadius, double sigma, double threshold, double minArea, double maxArea, boolean mergeAll, boolean watershedPostProcess, boolean excludeDAB, double cellExpansion, boolean limitExpansionByNucleusSize, boolean smoothBoundaries, boolean includeNuclei, boolean makeMeasurements) {
-		public void runDetection(double backgroundRadius, double maxBackground, double medianRadius, double sigma, double threshold, double minArea, double maxArea, boolean mergeAll, boolean watershedPostProcess, boolean excludeDAB, double cellExpansion, boolean smoothBoundaries, boolean includeNuclei, boolean makeMeasurements, int z, int t) {
+		public void runDetection(double backgroundRadius, double maxBackground, double medianRadius, double sigma, double threshold, double minArea, double maxArea, boolean mergeAll, boolean watershedPostProcess, boolean excludeDAB, double cellExpansion, boolean limitExpansionByNucleusSize, boolean smoothBoundaries, boolean includeNuclei, boolean makeMeasurements) {
 			
 			boolean updateNucleusROIs = rois == null || bpLoG == null;
 			updateNucleusROIs = updateNucleusROIs ? updateNucleusROIs : this.medianRadius != medianRadius;
 			this.medianRadius = medianRadius;
-			
-			updateNucleusROIs = updateNucleusROIs ? updateNucleusROIs : this.t != t || this.z != z;
-			this.z = z;
-			this.t = t;
 			
 			updateNucleusROIs = updateNucleusROIs ? updateNucleusROIs : this.backgroundRadius != backgroundRadius;
 			this.backgroundRadius = backgroundRadius;
@@ -982,24 +1078,29 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 			updateAnything = updateAnything ? updateAnything : this.makeMeasurements != makeMeasurements;
 			this.makeMeasurements = makeMeasurements;
 			
-//			updateAnything = updateAnything ? updateAnything : this.limitExpansionByNucleusSize != limitExpansionByNucleusSize;
-//			this.limitExpansionByNucleusSize = limitExpansionByNucleusSize;
+			updateAnything = updateAnything ? updateAnything : this.limitExpansionByNucleusSize != limitExpansionByNucleusSize;
+			this.limitExpansionByNucleusSize = limitExpansionByNucleusSize;
 			
 //			if (!updateAnything)
 //				return;
 			
-			doDetection(updateNucleusROIs);
-			
+			try {
+				doDetection(updateNucleusROIs);
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
 		}
 		
 		
 	}
+
 	
 	
 	@Override
 	public String getDescription() {
-		return "Default cell detection algorithm for brightfield images with nuclear or cytoplasmic staining";
+		return "Default cell detection algorithm for brightfield images with membrane staining";
 	}
+
 
 
 	@Override
@@ -1020,9 +1121,10 @@ public class WatershedCellDetection extends AbstractTileableDetectionPlugin<Buff
 		if (Double.isNaN(pxSize))
 			return params.getDoubleParameterValue("cellExpansion") > 0 ? 25 : 10;
 		double cellExpansion = params.getDoubleParameterValue("cellExpansionMicrons") / pxSize;
-		int overlap = cellExpansion > 0 ? (int)(cellExpansion + 10) : 10;
+		int overlap = cellExpansion > 0 ? (int)(cellExpansion * 2) : 10;
 //		System.out.println("Tile overlap: " + overlap + " pixels");
 		return overlap;
 	}
-		
+	
+	
 }
